Initial split into src/signer/incremental.rs.
