+++ src/signer/incremental.rs
@@ -1,0 +1,3218 @@
+#! /// Incremental signing.
+
+// Allow println! for now.
+#![allow(clippy::disallowed_macros)]
+
+use std::collections::{BTreeMap, HashMap, HashSet};
+use std::path::Path;
+use std::sync::{Arc, RwLock};
+use std::time::{Duration, UNIX_EPOCH};
+
+use bytes::{Bytes, BytesMut};
+use cascade_zonedata::{
+    LoadedZoneReader, OldParsedRecord, RegularRecord, SignedZonePatcher, SignedZoneReader,
+    SoaRecord,
+};
+use domain::base::iana::{Class, ZonemdAlgorithm, ZonemdScheme};
+use domain::base::name::FlattenInto;
+use domain::base::{
+    CanonicalOrd, Name, NameBuilder, Record, Rtype, Serial as DomainSerial, ToName, Ttl,
+};
+use domain::crypto::sign::SignRaw;
+use domain::dnssec::common::nsec3_hash;
+use domain::dnssec::sign::denial::nsec::{GenerateNsecConfig, generate_nsecs};
+use domain::dnssec::sign::denial::nsec3::{
+    GenerateNsec3Config, Nsec3ParamTtlMode, generate_nsec3s,
+};
+use domain::dnssec::sign::keys::SigningKey;
+use domain::dnssec::sign::keys::keyset::{KeyType, UnixTime};
+use domain::dnssec::sign::records::{DefaultSorter, RecordsIter, Rrset};
+use domain::dnssec::sign::signatures::rrsigs::sign_rrset;
+use domain::rdata::dnssec::{RtypeBitmap, Timestamp};
+use domain::rdata::nsec3::OwnerHash;
+use domain::rdata::{Nsec, Nsec3, Nsec3param, Soa, ZoneRecordData, Zonemd};
+use domain::utils::base32;
+use domain::zonefile::inplace::Entry;
+use domain::zonetree::StoredRecord;
+use domain_kmip::dep::kmip::client::pool::ConnectionManager;
+use domain_kmip::{self, ConnectionSettings, KeyUrl};
+use jiff::tz::TimeZone;
+use jiff::{Timestamp as JiffTimestamp, Zoned};
+use octseq::OctetsFrom;
+use rayon::slice::ParallelSliceMut;
+use tokio::time::Instant;
+use tracing::{debug, info, warn};
+use url::Url;
+
+use crate::center::Center;
+use crate::policy::{PolicyVersion, SignerDenialPolicy, SignerSerialPolicy};
+use crate::units::http_server::KmipServerState;
+use crate::units::key_manager::{KmipClientCredentialsFile, KmipServerCredentialsFileMode};
+use crate::units::zone_signer::{
+    KeyPair, KeySetState, PassThroughMode, SignerError, SigningStatusPerZone, ZoneSigner,
+    faketime_or_now,
+};
+use crate::zone::Zone;
+
+type Zrd = Record<Name<Bytes>, ZoneRecordData<Bytes, Name<Bytes>>>;
+type RtypeSet = HashSet<Rtype>;
+type ChangesValue = (RtypeSet, RtypeSet); // add set followed by delete set.
+
+pub struct WorkSpace<'a> {
+    pub keyset_state: KeySetState,
+    pub use_nsec3: bool,
+    pub verbose: bool,
+    pub policy: Arc<PolicyVersion>,
+    pub zone: Arc<Zone>,
+    pub center: Arc<Center>,
+    pub patch: SignedZonePatcher<'a>,
+
+    // Extra fields that should go to policy.
+    pub zonemd: HashSet<()>,
+    pub pass_through_mode: PassThroughMode,
+    // Extra fields that should go to state.
+}
+
+impl WorkSpace<'_> {
+    pub fn refresh_some_signatures(
+        &mut self,
+        iss: &mut IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        let effective_lifetime = Duration::from_secs(
+            (self.policy.signer.sig_validity_time - self.policy.signer.sig_remain_time) as u64,
+        );
+        let now = faketime_or_now();
+        let now_system_time = UNIX_EPOCH + Duration::from(now.clone());
+        let min_expire =
+            now_system_time + Duration::from_secs(self.policy.signer.sig_remain_time as u64);
+
+        let curr_last_signature_refresh = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.last_signature_refresh.clone()
+        };
+        let mut since_last_time: Duration = if now >= curr_last_signature_refresh {
+            <UnixTime as Into<Duration>>::into(now.clone())
+                - <UnixTime as Into<Duration>>::into(curr_last_signature_refresh.clone())
+        } else {
+            Duration::ZERO
+        };
+
+        // Limit to effective_lifetime in case of weird values.
+        if since_last_time > effective_lifetime {
+            since_last_time = effective_lifetime;
+        }
+
+        let total_signatures = iss.rrsigs.len();
+
+        let to_sign = since_last_time.as_secs_f64() * (total_signatures as f64)
+            / effective_lifetime.as_secs_f64();
+        let to_sign = to_sign.ceil() as usize;
+
+        // Collect expiration times, owner names, and types to figure out what
+        // to sign.
+        let mut expire_sigs = vec![];
+        for ((owner, rtype), r) in &iss.rrsigs {
+            let min_expiration = r
+                .iter()
+                .map(|r| {
+                    let ZoneRecordData::Rrsig(rrsig) = r.data() else {
+                        panic!("Rrsig expected");
+                    };
+                    rrsig.expiration().to_system_time(now_system_time)
+                })
+                .min()
+                .expect("minimum should exist");
+            let v = (min_expiration, owner, rtype);
+            expire_sigs.push(v);
+        }
+
+        expire_sigs.sort();
+
+        let mut new_sigs = vec![];
+        for (i, (expire, owner, rtype)) in expire_sigs.iter().enumerate() {
+            if *expire > min_expire && i >= to_sign {
+                break;
+            }
+
+            let key = ((*owner).clone(), **rtype);
+            if **rtype == Rtype::NSEC {
+                let record = iss.nsecs.get(&key.0).expect("NSEC record should exist");
+                let records = [record.clone()];
+                sign_records(
+                    &iss.origin,
+                    &records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            } else if **rtype == Rtype::NSEC3 {
+                let record = iss.nsec3s.get(&key.0).expect("NSEC3 record should exist");
+                let records = [record.clone()];
+                sign_records(
+                    &iss.origin,
+                    &records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            } else {
+                let records = iss.new_data.get(&key).expect("records should exist");
+                sign_records(
+                    &iss.origin,
+                    records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            };
+        }
+
+        for (sigs, rtype) in new_sigs {
+            let key = (sigs[0].owner().clone(), rtype);
+            iss.rrsigs.insert(key, sigs);
+        }
+
+        if to_sign != 0 {
+            // Only update last_signature_refresh when enough time has passed
+            // that at least one record got signed.
+            {
+                // Use a block to make sure that the mutex is clearly dropped.
+                let mut zone_state = self.zone.state.lock().unwrap();
+
+                zone_state.last_signature_refresh = now;
+                self.zone.mark_dirty(&mut zone_state, &self.center);
+            }
+        }
+        Ok(())
+    }
+
+    pub fn key_roll_signatures(
+        &mut self,
+        iss: &mut IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        let key_roll_time = Duration::from_secs(self.policy.signer.key_roll_time as u64);
+
+        let curr_key_roll = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.key_roll.clone()
+        };
+        let key_roll_start = curr_key_roll.as_ref().expect("should be there");
+
+        let now = faketime_or_now();
+
+        let since_start: Duration = <UnixTime as Into<Duration>>::into(now.clone())
+            - <UnixTime as Into<Duration>>::into(key_roll_start.clone());
+
+        let curr_key_tags = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.key_tags.clone()
+        };
+
+        if since_start > key_roll_time {
+            // Full roll. Make sure all signatures are made using the new keys.
+            // Clear key_roll when we are done.
+
+            let mut new_sigs = vec![];
+            for ((owner, rtype), r) in &iss.rrsigs {
+                let key_tags: HashSet<u16> = r
+                    .iter()
+                    .map(|r| {
+                        let ZoneRecordData::Rrsig(rrsig) = r.data() else {
+                            panic!("Rrsig expected");
+                        };
+                        rrsig.key_tag()
+                    })
+                    .collect();
+                if key_tags == curr_key_tags {
+                    // Nothing to do.
+                    continue;
+                }
+
+                let key = ((*owner).clone(), *rtype);
+                if *rtype == Rtype::NSEC {
+                    let record = iss.nsecs.get(&key.0).expect("NSEC record should exist");
+                    let records = [record.clone()];
+                    sign_records(
+                        &iss.origin,
+                        &records,
+                        &iss.keys,
+                        iss.inception,
+                        iss.expiration,
+                        &mut new_sigs,
+                    )?;
+                } else if *rtype == Rtype::NSEC3 {
+                    let record = iss.nsec3s.get(&key.0).expect("NSEC3 record should exist");
+                    let records = [record.clone()];
+                    sign_records(
+                        &iss.origin,
+                        &records,
+                        &iss.keys,
+                        iss.inception,
+                        iss.expiration,
+                        &mut new_sigs,
+                    )?;
+                } else {
+                    let records = iss.new_data.get(&key).expect("records should exist");
+                    sign_records(
+                        &iss.origin,
+                        records,
+                        &iss.keys,
+                        iss.inception,
+                        iss.expiration,
+                        &mut new_sigs,
+                    )?;
+                };
+            }
+
+            for (sigs, rtype) in new_sigs {
+                let key = (sigs[0].owner().clone(), rtype);
+                iss.rrsigs.insert(key, sigs);
+            }
+
+            // Clear key_roll.
+            {
+                // Use a block to make sure that the mutex is clearly dropped.
+                let mut zone_state = self.zone.state.lock().unwrap();
+
+                zone_state.key_roll = None;
+                self.zone.mark_dirty(&mut zone_state, &self.center);
+            }
+            return Ok(());
+        }
+
+        let total_signatures = iss.rrsigs.len();
+
+        let to_sign =
+            since_start.as_secs_f64() * (total_signatures as f64) / key_roll_time.as_secs_f64();
+        let to_sign = to_sign.ceil() as usize;
+
+        // owner names, types, and key tags to figure out what to sign.
+        let mut sigs_key_tags = vec![];
+        for ((owner, rtype), r) in &iss.rrsigs {
+            let key_tags: Vec<u16> = r
+                .iter()
+                .map(|r| {
+                    let ZoneRecordData::Rrsig(rrsig) = r.data() else {
+                        panic!("Rrsig expected");
+                    };
+                    rrsig.key_tag()
+                })
+                .collect();
+            let v = (owner, rtype, key_tags);
+            sigs_key_tags.push(v);
+        }
+
+        sigs_key_tags.sort();
+
+        let mut new_sigs = vec![];
+        for (i, (owner, rtype, key_tags)) in sigs_key_tags.iter().enumerate() {
+            if i >= to_sign {
+                break;
+            }
+
+            if HashSet::<u16>::from_iter(key_tags.iter().copied()) == curr_key_tags {
+                // Nothing to do.
+                continue;
+            }
+
+            let key = ((*owner).clone(), **rtype);
+            if **rtype == Rtype::NSEC {
+                let record = iss.nsecs.get(&key.0).expect("NSEC record should exist");
+                let records = [record.clone()];
+                sign_records(
+                    &iss.origin,
+                    &records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            } else if **rtype == Rtype::NSEC3 {
+                let record = iss.nsec3s.get(&key.0).expect("NSEC3 record should exist");
+                let records = [record.clone()];
+                sign_records(
+                    &iss.origin,
+                    &records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            } else {
+                let records = iss.new_data.get(&key).expect("records should exist");
+                sign_records(
+                    &iss.origin,
+                    records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            };
+        }
+
+        for (sigs, rtype) in new_sigs {
+            let key = (sigs[0].owner().clone(), rtype);
+            iss.rrsigs.insert(key, sigs);
+        }
+        Ok(())
+    }
+
+    pub fn handle_keyset_changed(&mut self) -> bool {
+        let mut apex_changed = false;
+
+        // Check the APEX RRtypes that need to be removed. We
+        // should get that from keyset, but currently we don't.
+        // Just have a fixed list.
+        let apex_remove: HashSet<Rtype> = [Rtype::DNSKEY, Rtype::CDS, Rtype::CDNSKEY].into();
+
+        let curr_apex_remove = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.apex_remove.clone()
+        };
+
+        if apex_remove != curr_apex_remove {
+            println!("APEX remove RRtypes changed: from {curr_apex_remove:?} to {apex_remove:?}",);
+
+            // Save the new apex_remove set.
+            {
+                // Use a block to make sure that the mutex is clearly dropped.
+                let mut zone_state = self.zone.state.lock().unwrap();
+
+                zone_state.apex_remove = apex_remove;
+                self.zone.mark_dirty(&mut zone_state, &self.center);
+            }
+            apex_changed = true;
+        }
+
+        // Check records that need to be added to the APEX.
+        let mut apex_extra = vec![];
+        apex_extra.extend_from_slice(&self.keyset_state.dnskey_rrset);
+        apex_extra.extend_from_slice(&self.keyset_state.cds_rrset);
+        apex_extra.sort();
+
+        let curr_apex_extra = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.apex_extra.clone()
+        };
+
+        if apex_extra != curr_apex_extra {
+            println!("APEX extra changed: from {curr_apex_extra:?} to {apex_extra:?}",);
+
+            // Save the new apex_extra list.
+            {
+                // Use a block to make sure that the mutex is clearly dropped.
+                let mut zone_state = self.zone.state.lock().unwrap();
+
+                zone_state.apex_extra = apex_extra;
+                self.zone.mark_dirty(&mut zone_state, &self.center);
+            }
+            apex_changed = true;
+        }
+
+        // Check if a ZSK/CSK roll has started.
+        let mut key_tags = HashSet::new();
+        for v in self.keyset_state.keyset.keys().values() {
+            let signer = match v.keytype() {
+                KeyType::Ksk(_) => false,
+                KeyType::Zsk(key_state) => key_state.signer(),
+                KeyType::Csk(_, key_state) => key_state.signer(),
+                KeyType::Include(_) => false,
+            };
+
+            if !signer {
+                continue;
+            }
+
+            key_tags.insert(v.key_tag());
+        }
+
+        let curr_key_tags = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.key_tags.clone()
+        };
+
+        if key_tags != curr_key_tags {
+            println!("key tags changed: from {curr_key_tags:?} to {key_tags:?}",);
+
+            // Save the new key tags set and key roll start time.
+            {
+                // Use a block to make sure that the mutex is clearly dropped.
+                let mut zone_state = self.zone.state.lock().unwrap();
+
+                zone_state.key_tags = key_tags;
+                zone_state.key_roll = Some(faketime_or_now());
+                self.zone.mark_dirty(&mut zone_state, &self.center);
+            }
+            apex_changed = true;
+        }
+        apex_changed
+    }
+
+    pub fn incremental_generate_diffs(
+        &mut self,
+        iss: &IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        /*
+                // apex records that were deleted.
+                for (k, old_rrs) in &iss.new_apex_saved {
+                    if let Some(new_rrs) = iss.new_apex.get(k) {
+                        if new_rrs == old_rrs {
+                            // No change.
+                            continue;
+                        }
+                        // Add the new records to a hash set and then check the old
+                        // ones against the set to see which ones are removed.
+                        let new_rrs: HashSet<&Zrd> = HashSet::from_iter(new_rrs.iter());
+                        for r in old_rrs {
+                            if new_rrs.contains(r) {
+                                continue;
+                            }
+                            let r: RegularRecord = r.clone().into();
+                            println!("apex patch.remove {r:?}");
+                            self.patch.remove(r).unwrap();
+                        }
+                    } else {
+                        for r in old_rrs {
+                            let r: RegularRecord = r.clone().into();
+                            println!("apex patch.remove {r:?}");
+                            self.patch.remove(r).unwrap();
+                        }
+                    }
+                }
+
+                // apex records that were added.
+                for (k, new_rrs) in &iss.new_apex {
+                    if let Some(old_rrs) = iss.new_apex_saved.get(k) {
+                        if new_rrs == old_rrs {
+                            // No change.
+                            continue;
+                        }
+                        // Add the old records to a hash set and then check the new
+                        // ones against the set to see which ones are added.
+                        let old_rrs: HashSet<&Zrd> = HashSet::from_iter(old_rrs.iter());
+                        for r in new_rrs {
+                            if old_rrs.contains(r) {
+                                continue;
+                            }
+                            let r: RegularRecord = r.clone().into();
+                            println!("apex patch.add {r:?}");
+                            self.patch.add(r).unwrap();
+                        }
+                    } else {
+                        for r in new_rrs {
+                            let r: RegularRecord = r.clone().into();
+                            println!("apex patch.add {r:?}");
+                            self.patch.add(r).unwrap();
+                        }
+                    }
+                }
+        */
+
+        // NSEC records that were deleted.
+        for (k, old_nsec) in &iss.old_nsecs {
+            if let Some(new_nsec) = iss.nsecs.get(k) {
+                if new_nsec == old_nsec {
+                    // No change.
+                    continue;
+                }
+                let old_nsec: RegularRecord = old_nsec.clone().into();
+                self.patch.remove(old_nsec).unwrap();
+            } else {
+                let old_nsec: RegularRecord = old_nsec.clone().into();
+                self.patch.remove(old_nsec).unwrap();
+            }
+        }
+
+        // NSEC records that were added.
+        for (k, new_nsec) in &iss.nsecs {
+            if let Some(old_nsec) = iss.old_nsecs.get(k) {
+                if new_nsec == old_nsec {
+                    // No change.
+                    continue;
+                }
+                let new_nsec: RegularRecord = new_nsec.clone().into();
+                self.patch.add(new_nsec).unwrap();
+            } else {
+                let new_nsec: RegularRecord = new_nsec.clone().into();
+                self.patch.add(new_nsec).unwrap();
+            }
+        }
+
+        // NSEC3 records that were deleted.
+        for (k, old_nsec3) in &iss.old_nsec3s {
+            if let Some(new_nsec3) = iss.nsec3s.get(k) {
+                if new_nsec3 == old_nsec3 {
+                    // No change.
+                    continue;
+                }
+                let old_nsec3: RegularRecord = old_nsec3.clone().into();
+                self.patch.remove(old_nsec3).unwrap();
+            } else {
+                let old_nsec3: RegularRecord = old_nsec3.clone().into();
+                self.patch.remove(old_nsec3).unwrap();
+            }
+        }
+
+        // NSEC3 records that were added.
+        for (k, new_nsec3) in &iss.nsec3s {
+            if let Some(old_nsec3) = iss.old_nsec3s.get(k) {
+                if new_nsec3 == old_nsec3 {
+                    // No change.
+                    continue;
+                }
+                let new_nsec3: RegularRecord = new_nsec3.clone().into();
+                self.patch.add(new_nsec3).unwrap();
+            } else {
+                let new_nsec3: RegularRecord = new_nsec3.clone().into();
+                self.patch.add(new_nsec3).unwrap();
+            }
+        }
+
+        // RRSIG records that were deleted.
+        for (k, old_rrsigs) in &iss.old_rrsigs {
+            if let Some(new_rrsigs) = iss.rrsigs.get(k) {
+                if new_rrsigs == old_rrsigs {
+                    // No change.
+                    continue;
+                }
+                // Add the new RRSIGs to a hash set and then check the old
+                // ones against the set to see which ones are removed.
+                let new_rrsigs: HashSet<&Zrd> = HashSet::from_iter(new_rrsigs.iter());
+                for r in old_rrsigs {
+                    if new_rrsigs.contains(r) {
+                        continue;
+                    }
+                    let r: RegularRecord = r.clone().into();
+                    //println!("patch.remove {r:?}");
+                    self.patch.remove(r).unwrap();
+                }
+            } else {
+                for r in old_rrsigs {
+                    let r: RegularRecord = r.clone().into();
+                    //println!("patch.remove {r:?}");
+                    self.patch.remove(r).unwrap();
+                }
+            }
+        }
+
+        // RRSIG records that were added.
+        for (k, new_rrsigs) in &iss.rrsigs {
+            if let Some(old_rrsigs) = iss.old_rrsigs.get(k) {
+                if new_rrsigs == old_rrsigs {
+                    // No change.
+                    continue;
+                }
+                // Add the old RRSIGs to a hash set and then check the new
+                // ones against the set to see which ones are added.
+                let old_rrsigs: HashSet<&Zrd> = HashSet::from_iter(old_rrsigs.iter());
+                for r in new_rrsigs {
+                    if old_rrsigs.contains(r) {
+                        continue;
+                    }
+                    let r: RegularRecord = r.clone().into();
+                    //println!("patch.add {r:?}");
+                    self.patch.add(r).unwrap();
+                }
+            } else {
+                for r in new_rrsigs {
+                    let r: RegularRecord = r.clone().into();
+                    //println!("patch.add {r:?}");
+                    self.patch.add(r).unwrap();
+                }
+            }
+        }
+
+        Ok(())
+        /*
+                let start = Instant::now();
+                let mut writer = {
+                    let filename = &self.config.zonefile_out;
+                    let file = File::create(filename)
+                        .map_err(|e| format!("unable to create file {}: {e}", filename.display()))?;
+                    BufWriter::new(file)
+                    // FileOrStdout::File(file)
+                };
+
+                for data in iss.new_data.values() {
+                    for rr in data {
+                        writer
+                            .write_fmt(format_args!("{}\n", rr.display_zonefile(DISPLAY_KIND)))
+                            .map_err(|e| format!("unable write signed zone: {e}"))?;
+                    }
+                }
+                for rr in iss.nsecs.values() {
+                    writer
+                        .write_fmt(format_args!("{}\n", rr.display_zonefile(DISPLAY_KIND)))
+                        .map_err(|e| format!("unable write signed zone: {e}"))?;
+                }
+                for rr in iss.nsec3s.values() {
+                    writer
+                        .write_fmt(format_args!("{}\n", rr.display_zonefile(DISPLAY_KIND)))
+                        .map_err(|e| format!("unable write signed zone: {e}"))?;
+                }
+                for data in iss.rrsigs.values() {
+                    for rr in data {
+                        let ZoneRecordData::Rrsig(rrsig) = rr.data() else {
+                            panic!("RRSIG expected");
+                        };
+                        let rr = Record::new(rr.owner(), rr.class(), rr.ttl(), YyyyMmDdHhMMSsRrsig(rrsig));
+                        writer
+                            .write_fmt(format_args!("{}\n", rr.display_zonefile(DISPLAY_KIND)))
+                            .map_err(|e| format!("unable write signed zone: {e}"))?;
+                    }
+                }
+                if self.verbose {
+                    println!("writing output took {:?}", start.elapsed());
+                }
+                Ok(())
+        */
+    }
+
+    /*
+        fn load_pass_through_dnskey(&mut self, iss: &mut IncrementalSigningState) -> Result<(), Error> {
+            // Assume that the APEX records have been copied from KeySetState to
+            // SignerState. Now update the APEX in new_data.
+
+            let mut dnskey_records = vec![];
+            let mut rrsig_records = vec![];
+
+            for r in &self.state.apex_extra {
+                let zonefile =
+                    domain::zonefile::inplace::Zonefile::from((r.to_string() + "\n").as_ref() as &str);
+                for entry in zonefile {
+                    let entry = entry.map_err::<Error, _>(|e| format!("bad entry: {e}\n").into())?;
+
+                    // We only care about records in a zonefile
+                    let Entry::Record(record) = entry else {
+                        continue;
+                    };
+
+                    if record.rtype() != Rtype::DNSKEY && record.rtype() != Rtype::RRSIG {
+                        continue;
+                    }
+
+                    let owner = record.owner().to_name::<Bytes>();
+                    let data = record.data().clone().try_flatten_into().unwrap();
+                    let r = Record::new(owner.clone(), record.class(), record.ttl(), data);
+
+                    if r.rtype() == Rtype::RRSIG {
+                        let ZoneRecordData::Rrsig(rrsig) = r.data() else {
+                            panic!("RRSIG expected");
+                        };
+                        if rrsig.type_covered() != Rtype::DNSKEY {
+                            continue;
+                        }
+                        rrsig_records.push(r);
+                    } else {
+                        dnskey_records.push(r);
+                    }
+                }
+            }
+
+            match self.config.pass_through_mode {
+                PassThroughMode::Off => unreachable!(),
+                PassThroughMode::CopyDnskeyRrset => {
+                    let key = (
+                        dnskey_records
+                            .first()
+                            .ok_or("at least one DNSKEY expected")?
+                            .owner()
+                            .clone(),
+                        Rtype::DNSKEY,
+                    );
+                    iss.new_data.insert(key.clone(), dnskey_records);
+                    iss.rrsigs.insert(key, rrsig_records);
+                }
+                PassThroughMode::MergeDnskeySignatures => {
+                    // Make sure the old and new DNSKEY RRsets are the same.
+                    let key = (iss.origin.clone(), Rtype::DNSKEY);
+                    let Some(old_dnskey_records) = iss.old_data.get(&key) else {
+                        return Err("A DNSKEY RRset should exist in the input zone".into());
+                    };
+                    let mut old_dnskey_records = old_dnskey_records.clone();
+                    old_dnskey_records.sort();
+                    dnskey_records.sort();
+                    if *old_dnskey_records != dnskey_records {
+                        return Err(
+                            "DNSKEY RRset in input has to be same as the DNSKEY RRset in keyset".into(),
+                        );
+                    }
+                    let Some(rrsigs) = iss.rrsigs.get(&key) else {
+                        return Err("RRSIGs expected for DNSKEY RRset".into());
+                    };
+                    let mut rrsigs = rrsigs.clone();
+                    rrsigs.append(&mut rrsig_records);
+                    iss.rrsigs.insert(key, rrsigs);
+                }
+            }
+
+            let key = (iss.origin.clone(), Rtype::ZONEMD);
+            if iss.new_data.contains_key(&key) {
+                return Err("Pass-through is not possible for zone input with ZONEMD".into());
+            }
+
+            Ok(())
+        }
+    */
+
+    pub fn add_zonemd(&self, _iss: &mut IncrementalSigningState) -> Result<(), SignerError> {
+        todo!();
+        /*
+                // Get the SOA record. We need that for the Serial and for the
+                // TTL.
+                let key = (iss.origin.clone(), Rtype::SOA);
+                let soa_records = iss
+                    .new_data
+                    .get(&key)
+                    .expect("SOA record should be present");
+                let ZoneRecordData::Soa(soa) = soa_records[0].data() else {
+                    panic!("SOA record expected");
+                };
+
+                let start = Instant::now();
+
+                // Create a Vec with all records to be able to sort them in canonical
+                // order. Ignore ZONEMD and RRSIGs of ZONEMD records.
+                let mut all = vec![];
+
+                let mut data: Vec<_> = iss
+                    .new_data
+                    .iter()
+                    .filter_map(|((o, t), r)| {
+                        if *o != iss.origin || *t != Rtype::ZONEMD {
+                            Some(r)
+                        } else {
+                            None
+                        }
+                    })
+                    .flatten()
+                    .collect();
+                all.append(&mut data);
+
+                let mut data: Vec<_> = iss.nsecs.values().collect();
+                all.append(&mut data);
+
+                let mut data: Vec<_> = iss.nsec3s.values().collect();
+                all.append(&mut data);
+
+                let mut data: Vec<_> = iss
+                    .rrsigs
+                    .iter()
+                    .filter_map(|((o, t), r)| {
+                        if *o != iss.origin || *t != Rtype::ZONEMD {
+                            Some(r)
+                        } else {
+                            None
+                        }
+                    })
+                    .flatten()
+                    .collect();
+                all.append(&mut data);
+
+                //all.sort_by(|e1, e2| CanonicalOrd::canonical_cmp(*e1, *e2));
+                all.par_sort_by(|e1, e2| CanonicalOrd::canonical_cmp(*e1, *e2));
+
+                if self.verbose {
+                    println!("ZONEMD prepare and sort took {:?}", start.elapsed());
+                }
+
+                let start = Instant::now();
+
+                let mut zonemd_records = vec![];
+                for z in &self.config.zonemd {
+                    if z.0 != ZonemdScheme::SIMPLE {
+                        return Err("unsupported zonemd scheme (only SIMPLE is supported)".into());
+                    }
+                    let mut buf: Vec<u8> = Vec::new();
+                    let mut ctx = match z.1 {
+                        ZonemdAlgorithm::SHA384 => digest::Context::new(&digest::SHA384),
+                        ZonemdAlgorithm::SHA512 => digest::Context::new(&digest::SHA512),
+                        _ => unreachable!(),
+                    };
+                    for r in &all {
+                        buf.clear();
+                        with_infallible(|| r.compose_canonical(&mut buf));
+                        ctx.update(&buf);
+                    }
+                    let digest = ctx.finish();
+                    let zonemd = Zonemd::new(
+                        soa.serial(),
+                        z.0,
+                        z.1,
+                        Bytes::copy_from_slice(digest.as_ref()),
+                    );
+                    let record = Record::new(
+                        iss.origin.clone(),
+                        soa_records[0].class(),
+                        soa_records[0].ttl(),
+                        ZoneRecordData::Zonemd(zonemd),
+                    );
+                    zonemd_records.push(record);
+                }
+
+                if self.verbose {
+                    println!("ZONEMD hash took {:?}", start.elapsed());
+                }
+
+                let key = (iss.origin.clone(), Rtype::ZONEMD);
+                let mut new_sigs = vec![];
+                sign_records(
+                    &iss.origin,
+                    &zonemd_records,
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+                iss.new_data.insert(key.clone(), zonemd_records);
+                iss.rrsigs.insert(key, new_sigs[0].0.clone());
+                Ok(())
+        */
+    }
+
+    fn update_soa_serial(&mut self, old_soa: &Zrd) -> Result<Zrd, SignerError> {
+        // Implement SOA serial policies. There are four policies:
+        // 1) Keep. Copy the serial from the unsigned zone. Refuse to sign
+        //    if the serial did not change.
+        // 2) Increment. Copy the serial from the unsigned zone but increment
+        //    the serial if the zone needs to be signed an the serial in
+        //    the unsigned zone did not change.
+        // 3) Unix timestamp. The current time in Unix seconds. Increment if
+        //    that does not result in a higher serial.
+        // 4) Broken down time (YYYYMMDDnn). The current day plus a serial
+        //    number. Implies increment to generate different serial numbers
+        //    over a day.
+
+        let ZoneRecordData::Soa(zone_soa) = old_soa.data() else {
+            unreachable!();
+        };
+
+        let curr_previous_serial = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.previous_serial
+        };
+
+        match self.policy.signer.serial_policy {
+            SignerSerialPolicy::Keep => {
+                if let Some(previous_serial) = curr_previous_serial
+                    && zone_soa.serial() <= previous_serial
+                {
+                    return Err(SignerError::SigningError(
+                        "Serial policy is Keep but upstream serial did not increase".to_string(),
+                    ));
+                }
+
+                // Save the new SOA serial.
+                {
+                    // Use a block to make sure that the mutex is clearly
+                    // dropped.
+                    let mut zone_state = self.zone.state.lock().unwrap();
+
+                    zone_state.previous_serial = Some(zone_soa.serial());
+                    self.zone.mark_dirty(&mut zone_state, &self.center);
+                }
+                Ok(old_soa.clone())
+            }
+            SignerSerialPolicy::Counter => {
+                // Always increment the serial number, ignore the serial
+                // number in the unsigned zone.
+                let previous_serial = if let Some(serial) = curr_previous_serial {
+                    serial
+                } else {
+                    DomainSerial::from(0)
+                };
+
+                let serial = previous_serial.add(1);
+
+                // Save the new SOA serial.
+                {
+                    // Use a block to make sure that the mutex is clearly
+                    // dropped.
+                    let mut zone_state = self.zone.state.lock().unwrap();
+
+                    zone_state.previous_serial = Some(serial);
+                    self.zone.mark_dirty(&mut zone_state, &self.center);
+                }
+
+                let new_soa = ZoneRecordData::Soa(Soa::new(
+                    zone_soa.mname().clone(),
+                    zone_soa.rname().clone(),
+                    serial,
+                    zone_soa.refresh(),
+                    zone_soa.retry(),
+                    zone_soa.expire(),
+                    zone_soa.minimum(),
+                ));
+                let record = Record::new(
+                    old_soa.owner().clone(),
+                    old_soa.class(),
+                    old_soa.ttl(),
+                    new_soa,
+                );
+
+                Ok(record)
+            }
+            SignerSerialPolicy::UnixTime => {
+                let mut serial = DomainSerial::now();
+                if let Some(previous_serial) = curr_previous_serial
+                    && serial <= previous_serial
+                {
+                    serial = previous_serial.add(1);
+                }
+
+                // Save the new SOA serial.
+                {
+                    // Use a block to make sure that the mutex is clearly
+                    // dropped.
+                    let mut zone_state = self.zone.state.lock().unwrap();
+
+                    zone_state.previous_serial = Some(serial);
+                    self.zone.mark_dirty(&mut zone_state, &self.center);
+                }
+
+                let new_soa = ZoneRecordData::Soa(Soa::new(
+                    zone_soa.mname().clone(),
+                    zone_soa.rname().clone(),
+                    serial,
+                    zone_soa.refresh(),
+                    zone_soa.retry(),
+                    zone_soa.expire(),
+                    zone_soa.minimum(),
+                ));
+
+                let record = Record::new(
+                    old_soa.owner().clone(),
+                    old_soa.class(),
+                    old_soa.ttl(),
+                    new_soa,
+                );
+
+                Ok(record)
+            }
+            SignerSerialPolicy::DateCounter => {
+                let ts = JiffTimestamp::now();
+                let zone = Zoned::new(ts, TimeZone::UTC);
+                let serial = ((zone.year() as u32 * 100 + zone.month() as u32) * 100
+                    + zone.day() as u32)
+                    * 100;
+                let mut serial: DomainSerial = serial.into();
+
+                if let Some(previous_serial) = curr_previous_serial
+                    && serial <= previous_serial
+                {
+                    serial = previous_serial.add(1);
+                }
+
+                // Save the new SOA serial.
+                {
+                    // Use a block to make sure that the mutex is clearly
+                    // dropped.
+                    let mut zone_state = self.zone.state.lock().unwrap();
+
+                    zone_state.previous_serial = Some(serial);
+                    self.zone.mark_dirty(&mut zone_state, &self.center);
+                }
+
+                let new_soa = ZoneRecordData::Soa(Soa::new(
+                    zone_soa.mname().clone(),
+                    zone_soa.rname().clone(),
+                    serial,
+                    zone_soa.refresh(),
+                    zone_soa.retry(),
+                    zone_soa.expire(),
+                    zone_soa.minimum(),
+                ));
+
+                let record = Record::new(
+                    old_soa.owner().clone(),
+                    old_soa.class(),
+                    old_soa.ttl(),
+                    new_soa,
+                );
+
+                Ok(record)
+            }
+        }
+    }
+
+    /*
+        fn run_notify_command(&self) -> Result<(), Error> {
+            if self.config.notify_command.is_empty() {
+                return Ok(()); // Nothing to do.
+            }
+
+            let output = Command::new(&self.config.notify_command[0])
+                .args(&self.config.notify_command[1..])
+                .output()
+                .map_err(|e| {
+                    format!(
+                        "unable to create new Command for {}: {e}",
+                        self.config.notify_command[0]
+                    )
+                })?;
+            if !output.status.success() {
+                println!("notify command failed with: {}", output.status);
+                io::stdout()
+                    .write_all(&output.stdout)
+                    .map_err(|e| format!("unable to write to stdout: {e}"))?;
+                io::stderr()
+                    .write_all(&output.stderr)
+                    .map_err(|e| format!("unable to write to stderr: {e}"))?;
+            }
+            Ok(())
+        }
+    */
+
+    pub fn sign_pass_through(&mut self) -> Result<(), SignerError> {
+        todo!();
+        /*
+                // Clear key_tags and key_roll to trigger resigning when
+                // pass-through mode is turned off. Also clear keyset_state_modified
+                // to trigger a reload of the keyset state when pass-through is
+                // turned off.
+                if !self.state.key_tags.is_empty() {
+                    self.state.key_tags = HashSet::new();
+                    self.state.keyset_state_modified = Timestamp::from(0).into();
+                    self.state_changed = true;
+                }
+                if self.state.key_roll.is_some() {
+                    self.state.key_roll = None;
+                    self.state_changed = true;
+                }
+
+                let mut iss = IncrementalSigningState::new(self)?;
+
+                let start = Instant::now();
+                load_signed_zone(&mut iss, &self.config.zonefile_in).unwrap();
+                if self.verbose {
+                    println!("loading signed zone took {:?}", start.elapsed());
+                }
+
+                // Re-use the signed data.
+                load_signed_only(&mut iss);
+
+                self.load_pass_through_dnskey(&mut iss)?;
+
+                self.incremental_write_output(&iss)?;
+                Ok(())
+        */
+    }
+
+    pub fn load_apex_records(
+        &mut self,
+        iss: &mut IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        // Assume that the APEX records have been copied from KeySetState to
+        // state. Now update the APEX in new_data.
+
+        // Delete all types in apex_remove.
+        let curr_apex_remove = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.apex_remove.clone()
+        };
+
+        for t in curr_apex_remove {
+            let key = (iss.origin.clone(), t);
+            iss.new_apex.remove(&t);
+            iss.rrsigs.remove(&key);
+        }
+
+        let curr_apex_extra = {
+            // Use a block to make sure that the mutex is clearly dropped.
+            let zone_state = self.zone.state.lock().unwrap();
+
+            zone_state.apex_extra.clone()
+        };
+
+        for r in curr_apex_extra {
+            let zonefile =
+                domain::zonefile::inplace::Zonefile::from((r.to_string() + "\n").as_ref() as &str);
+            for entry in zonefile {
+                let entry =
+                    entry.map_err(|e| SignerError::SigningError(format!("bad entry: {e}\n")))?;
+
+                // We only care about records in a zonefile
+                let Entry::Record(record) = entry else {
+                    continue;
+                };
+
+                let owner = record.owner().to_name::<Bytes>();
+                let data = record.data().clone().try_flatten_into().unwrap();
+                let r = Record::new(owner.clone(), record.class(), record.ttl(), data);
+
+                if r.rtype() == Rtype::RRSIG {
+                    let ZoneRecordData::Rrsig(rrsig) = r.data() else {
+                        panic!("RRSIG expected");
+                    };
+                    let key = (owner, rrsig.type_covered());
+                    let mut records = vec![r];
+                    if let Some(v) = iss.rrsigs.get_mut(&key) {
+                        v.append(&mut records);
+                    } else {
+                        iss.rrsigs.insert(key, records);
+                    }
+                } else {
+                    let key = r.rtype();
+                    let mut records = vec![r];
+                    if let Some(v) = iss.new_apex.get_mut(&key) {
+                        v.append(&mut records);
+                    } else {
+                        iss.new_apex.insert(key, records);
+                    }
+                }
+            }
+        }
+
+        if self.use_nsec3 {
+            // Copy the NSEC3PARAM record from the old_apex to the new_apex.
+            // The reason is that the NSEC3PARAM gets lost when the unsigned
+            // zone is loaded.
+            let nsec3param_records = iss
+                .old_apex
+                .get(&Rtype::NSEC3PARAM)
+                .expect("NSEC3PARAM should be present");
+            iss.new_apex
+                .insert(Rtype::NSEC3PARAM, nsec3param_records.to_vec());
+        }
+
+        if !self.zonemd.is_empty() {
+            let zonemd = Zonemd::new(
+                0.into(),
+                ZonemdScheme::SIMPLE,
+                ZonemdAlgorithm::SHA384,
+                Bytes::new(),
+            );
+            let record = Record::new(
+                iss.origin.clone(),
+                Class::IN,
+                Ttl::ZERO,
+                ZoneRecordData::Zonemd(zonemd),
+            );
+            let records = vec![record];
+            let key = (iss.origin.clone(), Rtype::ZONEMD);
+            iss.new_data.insert(key, records);
+        }
+
+        // Update the SOA serial.
+        let zone_soa_rr = &iss.new_apex.get(&Rtype::SOA).expect("SOA should exist")[0];
+        let new_soa = self.update_soa_serial(zone_soa_rr)?;
+        let new_rrset = vec![new_soa];
+        iss.new_apex.insert(Rtype::SOA, new_rrset);
+
+        let old_soa = iss.old_apex.get(&Rtype::SOA).unwrap();
+        for r in old_soa {
+            let r: SoaRecord = r.clone().into();
+            self.patch.remove_soa(r).unwrap();
+        }
+        let new_soa = iss.new_apex.get(&Rtype::SOA).unwrap();
+        for r in new_soa {
+            let r: SoaRecord = r.clone().into();
+            self.patch.add_soa(r).unwrap();
+        }
+
+        Ok(())
+    }
+
+    pub fn new_nsec_nsec3_sigs(
+        &self,
+        iss: &mut IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        let mut new_sigs = vec![];
+        if self.use_nsec3 {
+            for m in &iss.modified_nsecs {
+                let Some(nsec3) = iss.nsec3s.get(m) else {
+                    panic!("NSEC3 for {m} should exist");
+                };
+
+                let nsec3 = nsec3.clone();
+                sign_records(
+                    &iss.origin,
+                    &[nsec3],
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            }
+        } else {
+            for m in &iss.modified_nsecs {
+                let Some(nsec) = iss.nsecs.get(m) else {
+                    panic!("NSEC for {m} should exist");
+                };
+
+                let nsec = nsec.clone();
+                sign_records(
+                    &iss.origin,
+                    &[nsec],
+                    &iss.keys,
+                    iss.inception,
+                    iss.expiration,
+                    &mut new_sigs,
+                )?;
+            }
+        }
+        for (sig, rtype) in new_sigs {
+            let key = (sig[0].owner().clone(), rtype);
+            iss.rrsigs.insert(key, sig);
+        }
+        Ok(())
+    }
+
+    pub fn handle_nsec_nsec3(
+        &mut self,
+        iss: &mut IncrementalSigningState,
+    ) -> Result<(), SignerError> {
+        // Note that we could try to regenerate the NSEC(3). Assume that
+        // switching between NSEC, NSEC3, and NSEC3 opt-out (or other NSEC3
+        // parameter changes) is rare enough that we can just resign the full
+        // zone.
+        let opt_nsec3param = iss.old_apex.get(&Rtype::NSEC3PARAM);
+        if let Some(nsec3param_records) = opt_nsec3param {
+            // Zone was signed with NSEC3.
+            if !self.use_nsec3 {
+                // Zone is signed with NSEC3 but we want NSEC.
+                let start = Instant::now();
+                iss.remove_nsec_nsec3();
+                iss.new_nsec_chain()?;
+                println!("replacing NSEC3 with NSEC took {:?}", start.elapsed());
+                return Ok(());
+            }
+            let ZoneRecordData::Nsec3param(nsec3param) = nsec3param_records[0].data() else {
+                panic!("ZoneRecordData::Nsec3param expected");
+            };
+            if *nsec3param != iss.nsec3param {
+                // Parameters changed, resign.
+                let start = Instant::now();
+                iss.remove_nsec_nsec3();
+                iss.new_nsec3_chain()?;
+                if self.verbose {
+                    println!("updating NSEC3 parameters took {:?}", start.elapsed());
+                }
+                return Ok(());
+            }
+        } else {
+            // Zone was signed with NSEC, check if that is also the target.
+            if self.use_nsec3 {
+                // Resign the full zone with NSEC3.
+                let start = Instant::now();
+                iss.remove_nsec_nsec3();
+                iss.new_nsec3_chain()?;
+                println!("replacing NSEC with NSEC3 took {:?}", start.elapsed());
+                return Ok(());
+            }
+            // Stay with NSEC.
+        }
+        Ok(())
+    }
+}
+
+pub struct IncrementalSigningState {
+    origin: Name<Bytes>,
+    old_apex: HashMap<Rtype, Vec<Zrd>>,
+    new_apex: HashMap<Rtype, Vec<Zrd>>,
+    new_apex_saved: HashMap<Rtype, Vec<Zrd>>,
+    old_data: HashMap<(Name<Bytes>, Rtype), Vec<Zrd>>,
+    new_data: BTreeMap<(Name<Bytes>, Rtype), Vec<Zrd>>,
+    old_nsecs: BTreeMap<Name<Bytes>, Zrd>,
+    nsecs: BTreeMap<Name<Bytes>, Zrd>,
+    old_nsec3s: BTreeMap<Name<Bytes>, Zrd>,
+    nsec3s: BTreeMap<Name<Bytes>, Zrd>,
+    old_rrsigs: HashMap<(Name<Bytes>, Rtype), Vec<Zrd>>,
+    rrsigs: HashMap<(Name<Bytes>, Rtype), Vec<Zrd>>,
+
+    changes: HashMap<Name<Bytes>, ChangesValue>,
+    modified_nsecs: HashSet<Name<Bytes>>,
+    keys: Vec<SigningKey<Bytes, KeyPair>>,
+    inception: Timestamp,
+    expiration: Timestamp,
+
+    // NSEC3 paramters.
+    nsec3param: Nsec3param<Bytes>,
+}
+
+impl IncrementalSigningState {
+    pub fn new(
+        origin: Name<Bytes>,
+        policy: &PolicyVersion,
+        zone_signer: &ZoneSigner,
+        center: &Arc<Center>,
+        keyset_state: &KeySetState,
+        status: Arc<RwLock<SigningStatusPerZone>>,
+    ) -> Result<Self, SignerError> {
+        let keys = Self::load_keys(zone_signer, center, origin.clone(), keyset_state, status)?;
+
+        let now = faketime_or_now();
+        let now_u32 = Into::<Duration>::into(now.clone()).as_secs() as u32;
+        let inception = (now_u32 - policy.signer.sig_inception_offset).into();
+        let expiration = (now_u32 + policy.signer.sig_validity_time).into();
+
+        // This is the only way to deal with opt-out. There is no data type
+        // for flags or constant for opt-out. Creating an Nsec3param makes it
+        // possible to set opt-out.
+        let mut nsec3param = Nsec3param::default();
+        match &policy.signer.denial {
+            SignerDenialPolicy::NSec => (),
+            SignerDenialPolicy::NSec3 { opt_out } => {
+                if *opt_out {
+                    nsec3param.set_opt_out_flag();
+                }
+            }
+        }
+        Ok(Self {
+            origin,
+            old_apex: HashMap::new(),
+            new_apex: HashMap::new(),
+            new_apex_saved: HashMap::new(),
+            old_data: HashMap::new(),
+            new_data: BTreeMap::new(),
+            old_nsecs: BTreeMap::new(),
+            nsecs: BTreeMap::new(),
+            old_nsec3s: BTreeMap::new(),
+            nsec3s: BTreeMap::new(),
+            old_rrsigs: HashMap::new(),
+            rrsigs: HashMap::new(),
+            changes: HashMap::new(),
+            modified_nsecs: HashSet::new(),
+            keys,
+            inception,
+            expiration,
+            nsec3param,
+        })
+    }
+
+    fn load_keys(
+        zone_signer: &ZoneSigner,
+        center: &Arc<Center>,
+        zone_name: Name<Bytes>,
+        keyset_state: &KeySetState,
+        status: Arc<RwLock<SigningStatusPerZone>>,
+    ) -> Result<Vec<SigningKey<Bytes, KeyPair>>, SignerError> {
+        debug!("Loading dnst keyset signing keys");
+
+        let kmip_server_state_dir = &center.config.kmip_server_state_dir;
+        let kmip_credentials_store_path = &center.config.kmip_credentials_store_path;
+
+        debug!("Reading dnst keyset DNSKEY RRs and RRSIG RRs");
+        status.write().unwrap().current_action =
+            "Fetching apex RRs from the key manager".to_string();
+
+        // Read the DNSKEY RRs and DNSKEY RRSIG RR from the keyset state.
+
+        status.write().unwrap().current_action = "Loading signing keys".to_string();
+        // Load the signing keys indicated by the keyset state.
+        let mut signing_keys = vec![];
+        for (pub_key_name, key_info) in keyset_state.keyset.keys() {
+            // Only use active ZSKs or CSKs to sign the records in the zone.
+            if !matches!(key_info.keytype(),
+                KeyType::Zsk(key_state)|KeyType::Csk(_, key_state) if key_state.signer())
+            {
+                continue;
+            }
+
+            if let Some(priv_key_name) = key_info.privref() {
+                let priv_url = Url::parse(priv_key_name).expect("valid URL expected");
+                let pub_url = Url::parse(pub_key_name).expect("valid URL expected");
+
+                match (priv_url.scheme(), pub_url.scheme()) {
+                    ("file", "file") => {
+                        let priv_key_path = priv_url.path();
+                        debug!("Attempting to load private key '{priv_key_path}'.");
+
+                        let private_key = ZoneSigner::load_private_key(Path::new(priv_key_path))
+                            .map_err(|_| {
+                                SignerError::CannotReadPrivateKeyFile(priv_key_path.to_string())
+                            })?;
+
+                        let pub_key_path = pub_url.path();
+                        debug!("Attempting to load public key '{pub_key_path}'.");
+
+                        let public_key = ZoneSigner::load_public_key(Path::new(pub_key_path))
+                            .map_err(|_| {
+                                SignerError::CannotReadPublicKeyFile(pub_key_path.to_string())
+                            })?;
+
+                        let key_pair = domain::crypto::sign::KeyPair::from_bytes(
+                            &private_key,
+                            public_key.data(),
+                        )
+                        .map_err(|err| SignerError::InvalidKeyPairComponents(err.to_string()))?;
+                        let signing_key = SigningKey::new(
+                            zone_name.clone(),
+                            public_key.data().flags(),
+                            KeyPair::Domain(key_pair),
+                        );
+
+                        signing_keys.push(signing_key);
+                    }
+
+                    ("kmip", "kmip") => {
+                        let priv_key_url =
+                            KeyUrl::try_from(priv_url).map_err(SignerError::InvalidPublicKeyUrl)?;
+                        let pub_key_url =
+                            KeyUrl::try_from(pub_url).map_err(SignerError::InvalidPrivateKeyUrl)?;
+
+                        // TODO: Replace the connection pool if the persisted KMIP server settings
+                        // were updated more recently than the pool was created.
+
+                        let mut kmip_servers = zone_signer.kmip_servers.lock().unwrap();
+                        let kmip_conn_pool = match kmip_servers
+                            .entry(priv_key_url.server_id().to_string())
+                        {
+                            std::collections::hash_map::Entry::Occupied(e) => e.into_mut(),
+                            std::collections::hash_map::Entry::Vacant(e) => {
+                                // Try and load the KMIP server settings.
+                                let p = kmip_server_state_dir.join(priv_key_url.server_id());
+                                info!("Reading KMIP server state from '{p}'");
+                                let f = std::fs::File::open(p).unwrap();
+                                let kmip_server: KmipServerState =
+                                    serde_json::from_reader(f).unwrap();
+                                let KmipServerState {
+                                    server_id,
+                                    ip_host_or_fqdn: host,
+                                    port,
+                                    insecure,
+                                    connect_timeout,
+                                    read_timeout,
+                                    write_timeout,
+                                    max_response_bytes,
+                                    has_credentials,
+                                    ..
+                                } = kmip_server;
+
+                                let mut username = None;
+                                let mut password = None;
+                                if has_credentials {
+                                    let creds_file = KmipClientCredentialsFile::new(
+                                        kmip_credentials_store_path.as_std_path(),
+                                        KmipServerCredentialsFileMode::ReadOnly,
+                                    )
+                                    .unwrap();
+
+                                    let creds = creds_file.get(&server_id).ok_or(
+                                        SignerError::KmipServerCredentialsNeeded(server_id.clone()),
+                                    )?;
+
+                                    username = Some(creds.username.clone());
+                                    password = creds.password.clone();
+                                }
+
+                                let conn_settings = ConnectionSettings {
+                                    host,
+                                    port,
+                                    username,
+                                    password,
+                                    insecure,
+                                    client_cert: None, // TODO
+                                    server_cert: None, // TODO
+                                    ca_cert: None,     // TODO
+                                    connect_timeout: Some(connect_timeout),
+                                    read_timeout: Some(read_timeout),
+                                    write_timeout: Some(write_timeout),
+                                    max_response_bytes: Some(max_response_bytes),
+                                };
+
+                                let cloned_status = status.clone();
+                                let cloned_server_id = server_id.clone();
+                                tokio::task::spawn(async move {
+                                    cloned_status.write().unwrap().current_action =
+                                        format!("Connecting to KMIP server '{cloned_server_id}");
+                                });
+                                let pool = ConnectionManager::create_connection_pool(
+                                    server_id.clone(),
+                                    Arc::new(conn_settings.clone()),
+                                    10,
+                                    Some(Duration::from_secs(60)),
+                                    Some(Duration::from_secs(60)),
+                                )
+                                .map_err(|err| {
+                                    SignerError::CannotCreateKmipConnectionPool(server_id, err)
+                                })?;
+
+                                e.insert(pool)
+                            }
+                        };
+
+                        let _flags = priv_key_url.flags();
+
+                        let cloned_status = status.clone();
+                        let cloned_server_id = priv_key_url.server_id().to_string();
+                        tokio::task::spawn(async move {
+                            cloned_status.write().unwrap().current_action =
+                                format!("Fetching keys from KMIP server '{cloned_server_id}'");
+                        });
+
+                        let key_pair = KeyPair::Kmip(
+                            domain_kmip::sign::KeyPair::from_urls(
+                                priv_key_url,
+                                pub_key_url,
+                                kmip_conn_pool.clone(),
+                            )
+                            .map_err(|err| {
+                                SignerError::InvalidKeyPairComponents(err.to_string())
+                            })?,
+                        );
+
+                        let signing_key =
+                            SigningKey::new(zone_name.clone(), key_pair.dnskey().flags(), key_pair);
+
+                        signing_keys.push(signing_key);
+                    }
+
+                    (other1, other2) => {
+                        return Err(SignerError::InvalidKeyPairComponents(format!(
+                            "Using different key URI schemes ({other1} vs {other2}) for a public/private key pair is not supported."
+                        )));
+                    }
+                }
+
+                debug!("Loaded key pair for zone {zone_name} from key pair");
+            }
+        }
+
+        debug!("{} signing keys loaded", signing_keys.len());
+
+        // TODO: If signing is disabled for a zone should we then allow the
+        // unsigned zone to propagate through the pipeline?
+        if signing_keys.is_empty() {
+            warn!("No signing keys found for zone {zone_name}, aborting");
+            return Err(SignerError::SigningError(
+                "No signing keys found".to_string(),
+            ));
+        }
+
+        Ok(signing_keys)
+    }
+
+    pub fn load_signed_zone(
+        &mut self,
+        signed_reader: &SignedZoneReader,
+    ) -> Result<(), SignerError> {
+        // Collect records for a
+        // name/RRtype and store a complete RRset in a hash table.
+        let mut records = Vec::<Record<Name<Bytes>, ZoneRecordData<Bytes, Name<Bytes>>>>::new();
+        let mut rrsig_records = vec![];
+        let mut type_covered = Rtype::RRSIG;
+
+        for entry in signed_reader.all_records() {
+            let record: OldParsedRecord = entry.clone().into();
+            let record: StoredRecord = record.flatten_into();
+
+            match record.data() {
+                ZoneRecordData::Rrsig(rrsig) => {
+                    if rrsig_records.is_empty() {
+                        type_covered = rrsig.type_covered();
+                        rrsig_records.push(record);
+                        continue;
+                    }
+                    if record.owner() == rrsig_records[0].owner()
+                        && rrsig.type_covered() == type_covered
+                    {
+                        rrsig_records.push(record);
+                        continue;
+                    }
+
+                    let key = (rrsig_records[0].owner().clone(), type_covered);
+                    if let Some(v) = self.rrsigs.get_mut(&key) {
+                        v.append(&mut rrsig_records);
+                    } else {
+                        self.rrsigs.insert(key, rrsig_records);
+                    }
+                    type_covered = rrsig.type_covered();
+                    rrsig_records = vec![];
+                    rrsig_records.push(record);
+                }
+                ZoneRecordData::Nsec(_) => {
+                    // Assume (at most) one NSEC record per owner name.
+                    // Directly insert into the btree map.
+                    self.nsecs.insert(record.owner().clone(), record);
+                }
+                ZoneRecordData::Nsec3(_) => {
+                    // Assume (at most) one NSEC3 record per owner name.
+                    // Directly insert into the btree map.
+                    self.nsec3s.insert(record.owner().clone(), record);
+                }
+                _ => {
+                    if records.is_empty() {
+                        records.push(record);
+                        continue;
+                    }
+                    if record.owner() == records[0].owner() && record.rtype() == records[0].rtype()
+                    {
+                        records.push(record);
+                        continue;
+                    }
+                    let key = (records[0].owner().clone(), records[0].rtype());
+                    if key.0 == self.origin {
+                        if let Some(v) = self.old_apex.get_mut(&key.1) {
+                            v.append(&mut records);
+                        } else {
+                            self.old_apex.insert(key.1, records);
+                        }
+                    } else if let Some(v) = self.old_data.get_mut(&key) {
+                        v.append(&mut records);
+                    } else {
+                        self.old_data.insert(key, records);
+                    }
+                    records = vec![];
+                    records.push(record);
+                }
+            }
+        }
+
+        if !records.is_empty() {
+            let key = (records[0].owner().clone(), records[0].rtype());
+            if key.0 == self.origin {
+                if let Some(v) = self.old_apex.get_mut(&key.1) {
+                    v.append(&mut records);
+                } else {
+                    self.old_apex.insert(key.1, records);
+                }
+            } else if let Some(v) = self.old_data.get_mut(&key) {
+                v.append(&mut records);
+            } else {
+                self.old_data.insert(key, records);
+            }
+        }
+        if !rrsig_records.is_empty() {
+            let key = (rrsig_records[0].owner().clone(), type_covered);
+            if let Some(v) = self.rrsigs.get_mut(&key) {
+                v.append(&mut rrsig_records);
+            } else {
+                self.rrsigs.insert(key, rrsig_records);
+            }
+        }
+        self.old_nsecs = self.nsecs.clone();
+        self.old_nsec3s = self.nsec3s.clone();
+        self.old_rrsigs = self.rrsigs.clone();
+        Ok(())
+    }
+
+    pub fn load_unsigned_zone(&mut self, reader: &LoadedZoneReader) -> Result<(), SignerError> {
+        // Collect records for a
+        // name/RRtype and store a complete RRset in a btree.
+        let mut records = Vec::<Record<Name<Bytes>, ZoneRecordData<Bytes, Name<Bytes>>>>::new();
+
+        records.push(Into::<OldParsedRecord>::into(reader.soa().clone()).flatten_into());
+
+        for entry in reader.regular_records() {
+            let record: OldParsedRecord = entry.clone().into();
+            let record: StoredRecord = record.flatten_into();
+
+            let record: StoredRecord = record.flatten_into();
+
+            if records.is_empty() {
+                records.push(record);
+                continue;
+            }
+            if record.owner() == records[0].owner() && record.rtype() == records[0].rtype() {
+                records.push(record);
+                continue;
+            }
+            let key = (records[0].owner().clone(), records[0].rtype());
+            if key.0 == self.origin {
+                if let Some(v) = self.new_apex.get_mut(&key.1) {
+                    v.append(&mut records);
+                } else {
+                    self.new_apex.insert(key.1, records);
+                }
+            } else if let Some(v) = self.new_data.get_mut(&key) {
+                v.append(&mut records);
+            } else {
+                self.new_data.insert(key, records);
+            }
+            records = vec![];
+            records.push(record);
+        }
+
+        if !records.is_empty() {
+            let key = (records[0].owner().clone(), records[0].rtype());
+            if key.0 == self.origin {
+                if let Some(v) = self.new_apex.get_mut(&key.1) {
+                    v.append(&mut records);
+                } else {
+                    self.new_apex.insert(key.1, records);
+                }
+            } else if let Some(v) = self.new_data.get_mut(&key) {
+                v.append(&mut records);
+            } else {
+                self.new_data.insert(key, records);
+            }
+        }
+
+        // Save a copy of the loaded new_apex to createa diff later.
+        for (k, v) in &self.new_apex {
+            self.new_apex_saved.insert(*k, v.clone());
+        }
+
+        // Remove an NSEC3PARAM and ZONEMD that we got from the unsigned
+        // zone.
+        self.new_apex.remove(&Rtype::NSEC3PARAM);
+        self.new_apex.remove(&Rtype::ZONEMD);
+        Ok(())
+    }
+
+    pub fn load_signed_only(&mut self) {
+        // Copy old data to new data.
+
+        for (k, v) in &self.old_data {
+            self.new_data.insert(k.clone(), v.clone());
+        }
+        for (k, v) in &self.old_apex {
+            self.new_apex.insert(*k, v.clone());
+            self.new_apex_saved.insert(*k, v.clone());
+        }
+    }
+
+    pub fn initial_diffs(&mut self) -> Result<(), SignerError> {
+        let mut new_sigs = vec![];
+        for (_, new_rrset) in self.new_data.iter_mut() {
+            let key = (new_rrset[0].owner().clone(), new_rrset[0].rtype());
+            if let Some(mut old_rrset) = self.old_data.remove(&key) {
+                let rtype = new_rrset[0].rtype();
+                if (rtype == Rtype::DNSKEY || rtype == Rtype::CDS || rtype == Rtype::CDNSKEY)
+                    && *new_rrset[0].owner() == self.origin
+                {
+                    // At apex, these types are signed by the key manager. No
+                    // need to check for changes.
+                    continue;
+                }
+                old_rrset.sort_by(|a, b| a.as_ref().data().canonical_cmp(b.as_ref().data()));
+                new_rrset.sort_by(|a, b| a.as_ref().data().canonical_cmp(b.as_ref().data()));
+
+                if *old_rrset != *new_rrset && self.rrsigs.remove(&key).is_some() {
+                    sign_records(
+                        &self.origin,
+                        new_rrset,
+                        &self.keys,
+                        self.inception,
+                        self.expiration,
+                        &mut new_sigs,
+                    )?;
+                }
+            } else if let Some((added, _)) = self.changes.get_mut(&key.0) {
+                added.insert(new_rrset[0].rtype());
+            } else {
+                let mut added = HashSet::new();
+                let removed = HashSet::new();
+                added.insert(new_rrset[0].rtype());
+                self.changes.insert(key.0, (added, removed));
+            }
+        }
+        for (_, new_rrset) in self.new_apex.iter_mut() {
+            let key = (new_rrset[0].owner().clone(), new_rrset[0].rtype());
+            if let Some(mut old_rrset) = self.old_apex.remove(&key.1) {
+                let rtype = new_rrset[0].rtype();
+                if (rtype == Rtype::DNSKEY || rtype == Rtype::CDS || rtype == Rtype::CDNSKEY)
+                    && *new_rrset[0].owner() == self.origin
+                {
+                    // At apex, these types are signed by the key manager. No
+                    // need to check for changes.
+                    continue;
+                }
+                old_rrset.sort_by(|a, b| a.as_ref().data().canonical_cmp(b.as_ref().data()));
+                new_rrset.sort_by(|a, b| a.as_ref().data().canonical_cmp(b.as_ref().data()));
+
+                if *old_rrset != *new_rrset && self.rrsigs.remove(&key).is_some() {
+                    sign_records(
+                        &self.origin,
+                        new_rrset,
+                        &self.keys,
+                        self.inception,
+                        self.expiration,
+                        &mut new_sigs,
+                    )?;
+                }
+            } else if let Some((added, _)) = self.changes.get_mut(&key.0) {
+                added.insert(new_rrset[0].rtype());
+            } else {
+                let mut added = HashSet::new();
+                let removed = HashSet::new();
+                added.insert(new_rrset[0].rtype());
+                self.changes.insert(key.0, (added, removed));
+            }
+        }
+        for (sig, rtype) in new_sigs {
+            let key = (sig[0].owner().clone(), rtype);
+            self.rrsigs.insert(key, sig);
+        }
+        for old_rrset in self.old_data.values() {
+            // What is left in old_data is removed.
+            let rtype = old_rrset[0].rtype();
+            let key = (old_rrset[0].owner().clone(), rtype);
+
+            self.rrsigs.remove(&key);
+
+            if let Some((_, removed)) = self.changes.get_mut(&key.0) {
+                removed.insert(rtype);
+            } else {
+                let added = HashSet::new();
+                let mut removed = HashSet::new();
+                removed.insert(rtype);
+                self.changes.insert(key.0, (added, removed));
+            }
+        }
+        for old_rrset in self.old_apex.values() {
+            // What is left in old_data is removed.
+            let rtype = old_rrset[0].rtype();
+            let key = (old_rrset[0].owner().clone(), rtype);
+
+            self.rrsigs.remove(&key);
+
+            if let Some((_, removed)) = self.changes.get_mut(&key.0) {
+                removed.insert(rtype);
+            } else {
+                let added = HashSet::new();
+                let mut removed = HashSet::new();
+                removed.insert(rtype);
+                self.changes.insert(key.0, (added, removed));
+            }
+        }
+        Ok(())
+    }
+
+    pub fn incremental_nsec(&mut self) -> Result<(), SignerError> {
+        // Should changes be sorted or not? If changes is sorted we will
+        // process a new delegation before any glue. Which is more efficient.
+        // Otherwise if glue comes first, the glue will be signed and inserted
+        // in the NSEC chain only to be removed when the delegation is processed.
+        // However, we removing a delegation, the situation is reversed. For now
+        // assuming that sorting is not necessary.
+
+        let set_nsec_rrsig: HashSet<_> = [Rtype::NSEC, Rtype::RRSIG].into();
+
+        let changes = self.changes.clone();
+        for (key, (add, delete)) in &changes {
+            // The intersection between add and delete is empty.
+            assert!(add.intersection(delete).next().is_none());
+
+            if let Some(record_nsec) = self.nsecs.get(key) {
+                let record_nsec = record_nsec.clone();
+                let ZoneRecordData::Nsec(nsec) = record_nsec.data() else {
+                    panic!("NSEC record expected");
+                };
+
+                // Convert the existing RRtype bitmap into a hash set.
+                let mut curr = HashSet::new();
+                for rtype in nsec.types() {
+                    curr.insert(rtype);
+                }
+
+                // The intersection between curr and add is empty.
+                assert!(curr.intersection(add).next().is_none());
+
+                // delete is completely contained in curr. In other words the
+                // difference between delete and curr is empty.
+                assert!(delete.difference(&curr).next().is_none());
+
+                if add.contains(&Rtype::NS) {
+                    // Apex is special, but we can assume the NS RRset will not
+                    // be added to apex.
+                    assert!(*key != self.origin);
+
+                    // Remove the signatures for the existing types.
+                    for rtype in nsec.types().iter() {
+                        // When NS is added, we should keep the signatures for
+                        // DS and NSEC. The NSEC signature will be updated but
+                        // there is no point in removing it first. Do not try to
+                        // remove a signature for RRSIG because it does not exist.
+                        if rtype == Rtype::DS || rtype == Rtype::NSEC || rtype == Rtype::RRSIG {
+                            continue;
+                        }
+                        let key = (key.clone(), rtype);
+                        self.rrsigs.remove(&key);
+                    }
+
+                    // Restrict curr and add to these types.
+                    let mask: HashSet<Rtype> =
+                        [Rtype::NS, Rtype::DS, Rtype::NSEC, Rtype::RRSIG].into();
+
+                    let curr = curr.intersection(&mask).copied().collect();
+                    let add = add.intersection(&mask).copied().collect();
+
+                    // Update the NSEC record.
+                    nsec_update_bitmap(
+                        &record_nsec,
+                        nsec,
+                        &curr,
+                        &add,
+                        delete,
+                        &set_nsec_rrsig,
+                        self,
+                    );
+
+                    // Mark descendents as occluded after updating the bitmap.
+                    // The reason is that nsec_update_bitmap uses the current
+                    // next_name and nsec_set_occluded may change that.
+                    nsec_set_occluded(key, self);
+
+                    continue;
+                }
+                if delete.contains(&Rtype::NS) {
+                    // Apex is special, but we can assume the NS RRset will not
+                    // be removed from apex.
+                    assert!(*key != self.origin);
+
+                    // Curr does not include all types at this name. Add the
+                    // missing types to curr.
+                    let range_key = (key.clone(), 0.into());
+                    let range = self.new_data.range(range_key..);
+                    for ((r_name, r_type), _) in range {
+                        if r_name != key {
+                            break;
+                        }
+                        if add.contains(r_type) {
+                            // Skip what we are trying to add.
+                            continue;
+                        }
+                        curr.insert(*r_type);
+                    }
+
+                    let mut new = nsec_update_bitmap(
+                        &record_nsec,
+                        nsec,
+                        &curr,
+                        add,
+                        delete,
+                        &set_nsec_rrsig,
+                        self,
+                    );
+
+                    // Sign the types at this name except for NSEC, and RRSIG.
+                    new.remove(&Rtype::NSEC);
+                    new.remove(&Rtype::RRSIG);
+                    sign_rtype_set(key, &new, self)?;
+
+                    // Names that were previously occluded are no longer.
+                    nsec_clear_occluded(key, self)?;
+                    continue;
+                }
+                if *key != self.origin && nsec.types().contains(Rtype::NS) {
+                    // NS marks a delegation but only when the NS is not
+                    // at the apex.
+
+                    // If the add set contains DS then sign the DS RRset.
+                    if add.contains(&Rtype::DS) {
+                        let ds_set: HashSet<_> = [Rtype::DS].into();
+                        sign_rtype_set(key, &ds_set, self)?;
+                    }
+                    nsec_update_bitmap(
+                        &record_nsec,
+                        nsec,
+                        &curr,
+                        add,
+                        delete,
+                        &set_nsec_rrsig,
+                        self,
+                    );
+                    continue;
+                }
+
+                // The add types need to be signed.
+                sign_rtype_set(key, add, self)?;
+
+                nsec_update_bitmap(
+                    &record_nsec,
+                    nsec,
+                    &curr,
+                    add,
+                    delete,
+                    &set_nsec_rrsig,
+                    self,
+                );
+            } else {
+                if add.is_empty() {
+                    assert!(!delete.is_empty());
+                    // No need to do anything.
+                    continue;
+                }
+                assert!(delete.is_empty());
+                if is_occluded(key, self) {
+                    // No need to do anything.
+                    continue;
+                }
+
+                if add.contains(&Rtype::NS) {
+                    // Create a new NSEC record and sign only DS records (if any).
+                    let rtypebitmap = nsec_rtypebitmap_from_iterator(add.iter());
+                    nsec_insert(key, rtypebitmap, self);
+                    if add.contains(&Rtype::DS) {
+                        let ds_set: HashSet<_> = [Rtype::DS].into();
+                        sign_rtype_set(key, &ds_set, self)?;
+                    }
+
+                    // nsec_set_occluded expects the NSEC for key to exist.
+                    // So call this after inserting the new NSEC record.
+                    nsec_set_occluded(key, self);
+                    continue;
+                }
+                // Create a new NSEC record and sign all records.
+                let rtypebitmap = nsec_rtypebitmap_from_iterator(add.iter());
+                nsec_insert(key, rtypebitmap, self);
+                sign_rtype_set(key, add, self)?;
+            }
+        }
+        Ok(())
+    }
+
+    pub fn incremental_nsec3(&mut self) -> Result<(), SignerError> {
+        // Should changes be sorted or not? If changes is sorted we will
+        // process a new delegation before any glue. Which is more efficient.
+        // Otherwise if glue comes first, the glue will be signed and inserted
+        // in the NSEC chain only to be removed when the delegation is processed.
+        // However, we removing a delegation, the situation is reversed. For now
+        // assuming that sorting is not necessary.
+
+        let opt_out_flag = self.nsec3param.opt_out_flag();
+
+        let changes = self.changes.clone();
+        for (key, (add, delete)) in &changes {
+            // The intersection between add and delete is empty.
+            assert!(add.intersection(delete).next().is_none());
+
+            let (nsec3_hash_octets, nsec3_name) = nsec3_hash_parts(key, self);
+
+            if let Some(record_nsec3) = self.nsec3s.get(&nsec3_name) {
+                let record_nsec3 = record_nsec3.clone();
+                let ZoneRecordData::Nsec3(nsec3) = record_nsec3.data() else {
+                    panic!("NSEC3 record expected");
+                };
+
+                // Convert the existing RRtype bitmap into a hash set.
+                let mut curr = HashSet::new();
+                for rtype in nsec3.types() {
+                    curr.insert(rtype);
+                }
+
+                // The intersection between curr and add is empty.
+                assert!(curr.intersection(add).next().is_none());
+
+                // delete is completely contained in curr. In other words the
+                // difference between delete and curr is empty.
+                assert!(delete.difference(&curr).next().is_none());
+
+                if add.contains(&Rtype::NS) {
+                    // Apex is special, but we can assume the NS RRset will not
+                    // be added to apex.
+                    assert!(*key != self.origin);
+
+                    // Remove the signatures for the existing types.
+                    for rtype in nsec3.types().iter() {
+                        // When NS is added, we should keep the signatures for
+                        // DS. Do not try to remove a signature for RRSIG because
+                        // it does not exist.
+                        if rtype == Rtype::DS || rtype == Rtype::RRSIG {
+                            continue;
+                        }
+                        let key = (key.clone(), rtype);
+                        self.rrsigs.remove(&key);
+                    }
+
+                    // Restrict curr and add to these types.
+                    let mask: HashSet<Rtype> = [Rtype::NS, Rtype::DS, Rtype::RRSIG].into();
+
+                    let curr = curr.intersection(&mask).copied().collect();
+                    let add = add.intersection(&mask).copied().collect();
+
+                    // Update the NSEC3 record.
+                    nsec3_update_bitmap(key, &record_nsec3, nsec3, &curr, &add, delete, self);
+
+                    // Mark descendents as occluded after updating the bitmap.
+                    // The reason is that nsec3_update_bitmap uses the current
+                    // next_hash and nsec3_set_occluded may change that.
+                    nsec3_set_occluded(key, self);
+
+                    continue;
+                }
+                if delete.contains(&Rtype::NS) {
+                    // Apex is special, but we can assume the NS RRset will not
+                    // be removed from apex.
+                    assert!(*key != self.origin);
+
+                    // Curr does not include all types at this name. Add the
+                    // missing types to curr.
+                    let range_key = (key.clone(), 0.into());
+                    let range = self.new_data.range(range_key..);
+                    for ((r_name, r_type), _) in range {
+                        if r_name != key {
+                            break;
+                        }
+                        if add.contains(r_type) {
+                            // Skip what we are trying to add.
+                            continue;
+                        }
+                        curr.insert(*r_type);
+                    }
+
+                    let mut new =
+                        nsec3_update_bitmap(key, &record_nsec3, nsec3, &curr, add, delete, self);
+
+                    // Sign the types at this name except for NSEC, and RRSIG.
+                    new.remove(&Rtype::RRSIG);
+                    sign_rtype_set(key, &new, self)?;
+
+                    // Names that were previously occluded are no longer.
+                    nsec3_clear_occluded(key, self)?;
+                    continue;
+                }
+                if *key != self.origin && nsec3.types().contains(Rtype::NS) {
+                    // NS marks a delegation but only when the NS is not
+                    // at the apex.
+
+                    // If the add set contains DS then sign the DS RRset.
+                    if add.contains(&Rtype::DS) {
+                        let ds_set: HashSet<_> = [Rtype::DS].into();
+                        sign_rtype_set(key, &ds_set, self)?;
+                    }
+                    nsec3_update_bitmap(key, &record_nsec3, nsec3, &curr, add, delete, self);
+                    continue;
+                }
+
+                // The add types need to be signed.
+                sign_rtype_set(key, add, self)?;
+
+                nsec3_update_bitmap(key, &record_nsec3, nsec3, &curr, add, delete, self);
+            } else {
+                if add.is_empty() {
+                    assert!(!delete.is_empty());
+
+                    // Special magic for out-out. It is possible that an NS
+                    // record got deleted. With opt-out there will not be an
+                    // NSEC3 record if there is only a NS record and no DS record.
+                    if opt_out_flag && delete.contains(&Rtype::NS) {
+                        if is_occluded(key, self) {
+                            // No need to do anything.
+                            continue;
+                        }
+                        nsec3_clear_occluded(key, self)?;
+                        continue;
+                    }
+
+                    // No need to do anything.
+                    continue;
+                }
+                assert!(delete.is_empty());
+                if is_occluded(key, self) {
+                    // No need to do anything.
+                    continue;
+                }
+
+                // Just copy add in case we need to change it.
+                let mut add = add.clone();
+                if opt_out_flag {
+                    // We have a new record and no NSEC3 record exists. But in the
+                    // case of opt-out there may already be an NS record.
+                    // We are not at APEX because APEX always has an NSEC3
+                    // record.
+                    let tmpkey = (key.clone(), Rtype::NS);
+                    if self.new_data.contains_key(&tmpkey) {
+                        // Found an NS record. It is safe to add NS to the add
+                        // set.
+                        add.insert(Rtype::NS);
+                    }
+                }
+
+                if add.contains(&Rtype::NS) {
+                    if opt_out_flag {
+                        // Check if this is just an NS record. If so, don't
+                        // create an NSEC3 record.
+                        if !add.iter().any(|r| *r != Rtype::NS) {
+                            continue;
+                        }
+                    }
+                    // Create a new NSEC3 record and sign only DS records (if any).
+                    // If add contains DS then add RRSIG to add.
+
+                    let mut add = add.clone(); // In case we need to add RRSIG.
+                    if add.contains(&Rtype::DS) {
+                        let ds_set: HashSet<_> = [Rtype::DS].into();
+                        sign_rtype_set(key, &ds_set, self)?;
+                        add.insert(Rtype::RRSIG);
+                    }
+
+                    let rtypebitmap = nsec3_rtypebitmap_from_iterator(add.iter());
+
+                    nsec3_insert_full(key, nsec3_hash_octets, &nsec3_name, rtypebitmap, self);
+                    nsec3_set_occluded(key, self);
+                    continue;
+                }
+                // The new name is not a delegation. Add RRSIG to the set of
+                // Rtypes.
+                let mut add_with_rrsig = add.clone();
+                add_with_rrsig.insert(Rtype::RRSIG);
+
+                // Create a new NSEC3 record and sign all records.
+                let rtypebitmap = nsec3_rtypebitmap_from_iterator(add_with_rrsig.iter());
+                nsec3_insert_full(key, nsec3_hash_octets, &nsec3_name, rtypebitmap, self);
+                sign_rtype_set(key, &add, self)?;
+            }
+        }
+        Ok(())
+    }
+
+    fn remove_nsec_nsec3(&mut self) {
+        for k in self.nsecs.keys() {
+            let key = (k.clone(), Rtype::NSEC);
+            self.rrsigs.remove(&key);
+        }
+        self.nsecs = BTreeMap::new();
+
+        for k in self.nsec3s.keys() {
+            let key = (k.clone(), Rtype::NSEC3);
+            self.rrsigs.remove(&key);
+        }
+        self.nsec3s = BTreeMap::new();
+    }
+
+    fn new_nsec_chain(&mut self) -> Result<(), SignerError> {
+        let records = self.get_unsigned_sorted();
+        let records_iter = RecordsIter::new_from_refs(&records);
+        let config = GenerateNsecConfig::new();
+        let nsec_records = generate_nsecs(&self.origin, records_iter, &config)
+            .map_err(|e| SignerError::SigningError(format!("new_nsec_chain failed: {e}")))?;
+
+        // Collect signatures here.
+        let mut new_sigs = vec![];
+
+        for r in nsec_records {
+            let record = Record::new(
+                r.owner().clone(),
+                r.class(),
+                r.ttl(),
+                ZoneRecordData::Nsec(r.data().clone()),
+            );
+            self.nsecs.insert(record.owner().clone(), record.clone());
+            sign_records(
+                &self.origin,
+                &[record],
+                &self.keys,
+                self.inception,
+                self.expiration,
+                &mut new_sigs,
+            )?;
+        }
+        for (sig, rtype) in new_sigs {
+            let key = (sig[0].owner().clone(), rtype);
+            self.rrsigs.insert(key, sig);
+        }
+        Ok(())
+    }
+
+    fn new_nsec3_chain(&mut self) -> Result<(), SignerError> {
+        let records = self.get_unsigned_sorted();
+        let records_iter = RecordsIter::new_from_refs(&records);
+        let config = GenerateNsec3Config::<_, DefaultSorter>::new(self.nsec3param.clone())
+            .with_ttl_mode(Nsec3ParamTtlMode::SoaMinimum);
+        let nsec3_records = generate_nsec3s(&self.origin, records_iter, &config)
+            .map_err(|e| SignerError::SigningError(format!("generate_nsec3s failed: {e}")))?;
+
+        // Collect signatures here.
+        let mut new_sigs = vec![];
+
+        let r = nsec3_records.nsec3param;
+        let record = Record::new(
+            r.owner().clone(),
+            r.class(),
+            r.ttl(),
+            ZoneRecordData::Nsec3param(r.data().clone()),
+        );
+        let key = (record.owner().clone(), Rtype::NSEC3PARAM);
+        let records = vec![record.clone()];
+
+        // Insert in both old and new data.
+        sign_records(
+            &self.origin,
+            &[record],
+            &self.keys,
+            self.inception,
+            self.expiration,
+            &mut new_sigs,
+        )?;
+        self.old_data.insert(key.clone(), records.clone());
+        self.new_data.insert(key, records);
+
+        for r in nsec3_records.nsec3s {
+            let record = Record::new(
+                r.owner().clone(),
+                r.class(),
+                r.ttl(),
+                ZoneRecordData::Nsec3(r.data().clone()),
+            );
+            self.nsec3s.insert(record.owner().clone(), record.clone());
+            sign_records(
+                &self.origin,
+                &[record],
+                &self.keys,
+                self.inception,
+                self.expiration,
+                &mut new_sigs,
+            )?;
+        }
+        for (sig, rtype) in new_sigs {
+            let key = (sig[0].owner().clone(), rtype);
+            self.rrsigs.insert(key, sig);
+        }
+        Ok(())
+    }
+
+    fn get_unsigned_sorted(&self) -> Vec<&Zrd> {
+        // Create a Vec with all unsigned records to be able to sort them in
+        // canonical order.
+
+        let mut apex: Vec<_> = self.old_apex.values().flatten().collect();
+        let mut data: Vec<_> = self.old_data.values().flatten().collect();
+        data.append(&mut apex);
+        data.par_sort_by(|e1, e2| CanonicalOrd::canonical_cmp(*e1, *e2));
+
+        data
+    }
+}
+
+fn sign_records(
+    origin: &Name<Bytes>,
+    records: &[Zrd],
+    keys: &[SigningKey<Bytes, KeyPair>],
+    inception: Timestamp,
+    expiration: Timestamp,
+    new_sigs: &mut Vec<(Vec<Zrd>, Rtype)>,
+) -> Result<(), SignerError> {
+    let rtype = records[0].rtype();
+    if (rtype == Rtype::DNSKEY || rtype == Rtype::CDS || rtype == Rtype::CDNSKEY)
+        && records[0].owner() == origin
+    {
+        // These records get signed with the KSK(s). Don't touch
+        // the signatures.
+        return Ok(());
+    }
+
+    let rrset = Rrset::new_from_owned(records)
+        .map_err(|e| SignerError::SigningError(format!("Rrset::new failed: {e}")))?;
+    let mut rrsig_records = vec![];
+    for key in keys {
+        let rrsig = sign_rrset(key, &rrset, inception, expiration)
+            .map_err(|e| SignerError::SigningError(format!("signing failed: {e}")))?;
+        let record = Record::new(
+            rrsig.owner().clone(),
+            rrsig.class(),
+            rrsig.ttl(),
+            ZoneRecordData::Rrsig(rrsig.data().clone()),
+        );
+        rrsig_records.push(record);
+    }
+    new_sigs.push((rrsig_records, rrset.rtype()));
+    Ok(())
+}
+
+fn nsec_insert(
+    name: &Name<Bytes>,
+    rtypebitmap: RtypeBitmap<Bytes>,
+    iss: &mut IncrementalSigningState,
+) {
+    // Try to find the NSEC record that comes before the one we are trying
+    // to insert. Assume that the APEX NSEC will always exist can sort
+    // before anything else.
+    let mut range = iss.nsecs.range::<Name<_>, _>(..name);
+    let (previous_name, previous_record) = range
+        .next_back()
+        .expect("previous NSEC record should exist");
+    let previous_name = previous_name.clone();
+    let previous_record = previous_record.clone();
+    drop(range);
+    let ZoneRecordData::Nsec(previous_nsec) = previous_record.data() else {
+        panic!("NSEC record expected");
+    };
+    let next = previous_nsec.next_name();
+    let new_nsec = Nsec::new(next.clone(), rtypebitmap);
+    let new_record = Record::new(
+        name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec(new_nsec),
+    );
+    iss.nsecs.insert(name.clone(), new_record);
+    iss.modified_nsecs.insert(name.clone());
+    let previous_nsec = Nsec::new(name.clone(), previous_nsec.types().clone());
+    let previous_record = Record::new(
+        previous_name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec(previous_nsec),
+    );
+    iss.nsecs.insert(previous_name.clone(), previous_record);
+    iss.modified_nsecs.insert(previous_name.clone());
+}
+
+fn nsec_remove(name: &Name<Bytes>, next_name: &Name<Bytes>, iss: &mut IncrementalSigningState) {
+    // Try to find the NSEC record that comes before the one we are trying
+    // to remove. Assume that the APEX NSEC will always exist can sort
+    // before anything else.
+    let mut range = iss.nsecs.range::<Name<_>, _>(..name);
+    let (previous_name, previous_record) = range
+        .next_back()
+        .expect("previous NSEC record should exist");
+    let previous_name = previous_name.clone();
+    let previous_record = previous_record.clone();
+    drop(range);
+    let ZoneRecordData::Nsec(previous_nsec) = previous_record.data() else {
+        panic!("NSEC record expected");
+    };
+    let previous_nsec = Nsec::new(next_name.clone(), previous_nsec.types().clone());
+    let previous_record = Record::new(
+        previous_name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec(previous_nsec),
+    );
+    iss.nsecs.insert(previous_name.clone(), previous_record);
+    iss.modified_nsecs.insert(previous_name.clone());
+    iss.nsecs.remove(name);
+    iss.modified_nsecs.remove(name);
+    let key = (name.clone(), Rtype::NSEC);
+    iss.rrsigs.remove(&key);
+}
+
+// Return the effective result HashSet even when the NSEC record gets deleted.
+fn nsec_update_bitmap(
+    record: &Zrd,
+    nsec: &Nsec<Bytes, Name<Bytes>>,
+    curr: &HashSet<Rtype>,
+    add: &HashSet<Rtype>,
+    delete: &HashSet<Rtype>,
+    set_nsec_rrsig: &HashSet<Rtype>,
+    iss: &mut IncrementalSigningState,
+) -> HashSet<Rtype> {
+    // Update curr.
+    let curr: HashSet<_> = curr.union(add).copied().collect();
+    let curr = curr.difference(delete).copied().collect();
+
+    let owner = record.owner();
+    if curr == *set_nsec_rrsig {
+        nsec_remove(owner, nsec.next_name(), iss);
+        return curr;
+    }
+
+    let rtypebitmap = nsec_rtypebitmap_from_iterator(curr.iter());
+    let nsec = Nsec::new(nsec.next_name().clone(), rtypebitmap);
+    let record = Record::new(
+        record.owner().clone(),
+        record.class(),
+        record.ttl(),
+        ZoneRecordData::Nsec(nsec),
+    );
+    iss.nsecs.insert(owner.clone(), record);
+
+    iss.modified_nsecs.insert(owner.clone());
+    curr
+}
+
+fn nsec_set_occluded(name: &Name<Bytes>, iss: &mut IncrementalSigningState) {
+    let Some(nsec_record) = iss.nsecs.get(name) else {
+        panic!("NSEC for {name} expected to exist");
+    };
+    let ZoneRecordData::Nsec(nsec) = nsec_record.data() else {
+        panic!("NSEC record expected");
+    };
+    let nsec = nsec.clone();
+    let mut next = nsec.next_name().clone();
+    loop {
+        if !next.ends_with(name) {
+            break;
+        }
+
+        // For consistency, make sure next is not equal to name.
+        if next == name {
+            break;
+        }
+        let curr = next;
+        let Some(nsec_record) = iss.nsecs.get(&curr) else {
+            panic!("NSEC for {name} expected to exist");
+        };
+        let ZoneRecordData::Nsec(nsec) = nsec_record.data() else {
+            panic!("NSEC record expected");
+        };
+        let nsec = nsec.clone();
+        next = nsec.next_name().clone();
+
+        nsec_remove(&curr, &next, iss);
+
+        // Remove all signatures.
+        for rtype in nsec.types().iter() {
+            let key = (curr.clone(), rtype);
+            iss.rrsigs.remove(&key);
+        }
+    }
+}
+
+fn nsec_clear_occluded(
+    name: &Name<Bytes>,
+    iss: &mut IncrementalSigningState,
+) -> Result<(), SignerError> {
+    let key = (name.clone(), Rtype::SOA);
+    let range = iss.new_data.range(key..);
+    let mut opt_curr_name: Option<&Name<Bytes>> = None;
+    let mut curr_types: HashSet<Rtype> = HashSet::new();
+    let mut work = vec![];
+
+    // Keep track of delegations. Name below a delegation remain occluded.
+    let mut delegation: Option<Name<Bytes>> = None;
+
+    for ((key_name, key_rtype), _) in range {
+        // There is no easy way to avoid name showing up in the range. Just
+        // filter out name.
+        if key_name == name {
+            continue;
+        }
+
+        // Make sure curr_name is below name.
+        if !key_name.ends_with(name) {
+            break;
+        }
+        if let Some(d) = &delegation
+            && key_name.ends_with(d)
+            && key_name != d
+        {
+            // Skip.
+            continue;
+        }
+
+        if *key_rtype == Rtype::NS {
+            // Set key_name as a delegation.
+            delegation = Some(key_name.clone());
+        }
+        if let Some(curr_name) = opt_curr_name {
+            if key_name == curr_name {
+                curr_types.insert(*key_rtype);
+            } else {
+                work.push((curr_name.clone(), curr_types));
+                opt_curr_name = Some(key_name);
+                curr_types = [*key_rtype].into();
+            }
+        } else {
+            opt_curr_name = Some(key_name);
+            curr_types.insert(*key_rtype);
+        }
+    }
+    if let Some(curr_name) = opt_curr_name {
+        work.push((curr_name.clone(), curr_types));
+    }
+    for (curr_name, curr_types) in work {
+        let mut curr_types = if curr_types.contains(&Rtype::NS) {
+            let has_ds = curr_types.contains(&Rtype::DS);
+            let mut curr_types: HashSet<Rtype> = [Rtype::NS].into();
+            if has_ds {
+                curr_types.insert(Rtype::DS);
+            }
+            curr_types
+        } else {
+            curr_types
+        };
+        let rtypebitmap = nsec_rtypebitmap_from_iterator(curr_types.iter());
+
+        // Make sure NS doesn't get signed.
+        curr_types.remove(&Rtype::NS);
+        sign_rtype_set(&curr_name, &curr_types, iss)?;
+        nsec_insert(&curr_name, rtypebitmap, iss);
+    }
+    Ok(())
+}
+
+fn nsec_rtypebitmap_from_iterator<'a, I>(iter: I) -> RtypeBitmap<Bytes>
+where
+    I: Iterator<Item = &'a Rtype>,
+{
+    let mut rtypebitmap = RtypeBitmap::<Bytes>::builder();
+    rtypebitmap.add(Rtype::NSEC).expect("should not fail");
+    rtypebitmap.add(Rtype::RRSIG).expect("should not fail");
+    for rtype in iter {
+        rtypebitmap.add(*rtype).expect("should not fail");
+    }
+    rtypebitmap.finalize()
+}
+
+fn nsec3_update(
+    owner: &Name<Bytes>,
+    nsec3_record: &Zrd,
+    nsec3: &Nsec3<Bytes>,
+    rtypes: &HashSet<Rtype>,
+    iss: &mut IncrementalSigningState,
+) {
+    // Just update an NSEC3 record without further logic.
+    let rtypebitmap = nsec3_rtypebitmap_from_iterator(rtypes.iter());
+    let nsec3 = Nsec3::new(
+        iss.nsec3param.hash_algorithm(),
+        iss.nsec3param.flags(),
+        iss.nsec3param.iterations(),
+        iss.nsec3param.salt().clone(),
+        nsec3.next_owner().clone(),
+        rtypebitmap,
+    );
+    let record = Record::new(
+        nsec3_record.owner().clone(),
+        nsec3_record.class(),
+        nsec3_record.ttl(),
+        ZoneRecordData::Nsec3(nsec3),
+    );
+    iss.nsec3s.insert(owner.clone(), record);
+
+    iss.modified_nsecs.insert(owner.clone());
+}
+
+fn nsec3_remove_full(
+    name: &Name<Bytes>,
+    nsec3_name: &Name<Bytes>,
+    nsec3_next: &OwnerHash<Bytes>,
+    iss: &mut IncrementalSigningState,
+) {
+    nsec3_remove_one(nsec3_name, nsec3_next, iss);
+
+    // Assume that we never remove the APEX. So the parent always exists.
+    let name = name.parent().expect("should exist");
+    nsec3_remove_et(&name, iss);
+}
+
+fn nsec3_remove_et(name: &Name<Bytes>, iss: &mut IncrementalSigningState) {
+    // Check if name is an ET. If so remove it and see if the parent is
+    // also an ET.
+    //
+    // Take a simple approach to check if a name is an ET: first lookup
+    // the NSEC3 record for name and check that the bitmap is empty. Then
+    // check all descendent names and check that none of them has an
+    // NSEC3 record.
+    let mut name = name.clone();
+    loop {
+        if !name.ends_with(&iss.origin) {
+            // This is weird, we should never be able to get beyond APEX.
+            // Just ignore this.
+            return;
+        }
+        if name == iss.origin {
+            // Never remove the NSEC3 record for APEX.
+            return;
+        }
+
+        let (_, nsec3_name) = nsec3_hash_parts(&name, iss);
+
+        let Some(record_nsec3) = iss.nsec3s.get(&nsec3_name) else {
+            // No NSEC3 record, nothing to do.
+            return;
+        };
+
+        let ZoneRecordData::Nsec3(nsec3) = record_nsec3.data() else {
+            panic!("NSEC3 record expected");
+        };
+
+        if !nsec3.types().is_empty() {
+            // There are types here.
+            return;
+        }
+
+        // Check the descendents.
+        let key = (name.clone(), Rtype::SOA);
+        let range = iss.new_data.range(key..);
+        let mut opt_curr_name: Option<&Name<Bytes>> = None;
+
+        for ((key_name, _), _) in range {
+            // There is no easy way to avoid name showing up in the range. Just
+            // filter out name.
+            if *key_name == name {
+                continue;
+            }
+
+            // Make sure curr_name is below name.
+            if !key_name.ends_with(&name) {
+                break;
+            }
+
+            if let Some(curr_name) = opt_curr_name
+                && key_name == curr_name
+            {
+                // Already checked.
+                continue;
+            }
+
+            opt_curr_name = Some(key_name);
+
+            let (_, nsec3_name) = nsec3_hash_parts(key_name, iss);
+
+            if iss.nsec3s.contains_key(&nsec3_name) {
+                // NSEC3 record is found. Our target is not an ET.
+                return;
+            };
+        }
+
+        // No descendents with NSEC3 records are found. Delete this one.
+        let next_owner = nsec3.next_owner().clone();
+        nsec3_remove_one(&nsec3_name, &next_owner, iss);
+
+        // We remove the NSEC3 record for the name. Get the parent. We should
+        // be below APEX, so the parent has to exist.
+        name = name.parent().expect("parent should exist");
+    }
+}
+
+fn nsec3_remove_one(
+    nsec3_name: &Name<Bytes>,
+    nsec3_next: &OwnerHash<Bytes>,
+    iss: &mut IncrementalSigningState,
+) {
+    // Try to find the NSEC3 record that comes before the one we are trying
+    // to remove.
+    let mut range = iss.nsec3s.range::<Name<_>, _>(..nsec3_name);
+    let (previous_name, previous_record) = if let Some(kv) = range.next_back() {
+        kv
+    } else {
+        let mut range = iss.nsec3s.range::<Name<_>, _>(nsec3_name..);
+        range
+            .next_back()
+            .expect("at least one element should exist")
+    };
+
+    let previous_name = previous_name.clone();
+    let previous_record = previous_record.clone();
+    drop(range);
+    let ZoneRecordData::Nsec3(previous_nsec) = previous_record.data() else {
+        panic!("NSEC3 record expected");
+    };
+    let previous_nsec3 = Nsec3::new(
+        iss.nsec3param.hash_algorithm(),
+        iss.nsec3param.flags(),
+        iss.nsec3param.iterations(),
+        iss.nsec3param.salt().clone(),
+        nsec3_next.clone(),
+        previous_nsec.types().clone(),
+    );
+    let previous_record = Record::new(
+        previous_name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec3(previous_nsec3),
+    );
+    iss.nsec3s.insert(previous_name.clone(), previous_record);
+    iss.modified_nsecs.insert(previous_name.clone());
+    iss.nsec3s.remove(nsec3_name);
+    iss.modified_nsecs.remove(nsec3_name);
+    let key = (nsec3_name.clone(), Rtype::NSEC3);
+    iss.rrsigs.remove(&key);
+}
+
+fn nsec3_set_occluded(name: &Name<Bytes>, iss: &mut IncrementalSigningState) {
+    // Loop over all names below name, if there is an NSEC3 record then
+    // delete all signatures and the NSEC3 record.
+
+    let key = (name.clone(), Rtype::SOA);
+    let range = iss.new_data.range(key..);
+    let mut opt_curr_name: Option<&Name<Bytes>> = None;
+    let mut work = vec![];
+
+    for ((key_name, _), _) in range {
+        // There is no easy way to avoid name showing up in the range. Just
+        // filter out name.
+        if key_name == name {
+            continue;
+        }
+
+        // Make sure curr_name is below name.
+        if !key_name.ends_with(name) {
+            break;
+        }
+
+        if let Some(curr_name) = opt_curr_name
+            && key_name == curr_name
+        {
+            // Looked at this name already.
+            continue;
+        }
+
+        opt_curr_name = Some(key_name);
+
+        let (_, nsec3_name) = nsec3_hash_parts(key_name, iss);
+
+        let Some(record_nsec3) = iss.nsec3s.get(&nsec3_name) else {
+            // No NSEC3 record, nothing to do.
+            continue;
+        };
+
+        let ZoneRecordData::Nsec3(nsec3) = record_nsec3.data() else {
+            panic!("NSEC3 record expected");
+        };
+
+        work.push((key_name.clone(), nsec3_name));
+
+        // Remove all signatures.
+        for rtype in nsec3.types().iter() {
+            let key = (key_name.clone(), rtype);
+            iss.rrsigs.remove(&key);
+        }
+    }
+    for (key_name, nsec3_name) in work {
+        let record_nsec3 = iss.nsec3s.get(&nsec3_name).expect("NSEC3 should exist");
+
+        let ZoneRecordData::Nsec3(nsec3) = record_nsec3.data() else {
+            panic!("NSEC3 record expected");
+        };
+
+        let nsec3_next = nsec3.next_owner().clone();
+        nsec3_remove_full(&key_name, &nsec3_name, &nsec3_next, iss);
+    }
+}
+
+fn nsec3_clear_occluded(
+    name: &Name<Bytes>,
+    iss: &mut IncrementalSigningState,
+) -> Result<(), SignerError> {
+    let key = (name.clone(), Rtype::SOA);
+    let range = iss.new_data.range(key..);
+    let mut opt_curr_name: Option<&Name<Bytes>> = None;
+    let mut curr_types: HashSet<Rtype> = HashSet::new();
+    let mut work = vec![];
+
+    // Keep track of delegations. Name below a delegation remain occluded.
+    let mut delegation: Option<Name<Bytes>> = None;
+
+    for ((key_name, key_rtype), _) in range {
+        // There is no easy way to avoid name showing up in the range. Just
+        // filter out name.
+        if key_name == name {
+            continue;
+        }
+
+        // Make sure curr_name is below name.
+        if !key_name.ends_with(name) {
+            break;
+        }
+        if let Some(d) = &delegation
+            && key_name.ends_with(d)
+            && key_name != d
+        {
+            // Skip.
+            continue;
+        }
+
+        if *key_rtype == Rtype::NS {
+            // Set key_name as a delegation.
+            delegation = Some(key_name.clone());
+        }
+        if let Some(curr_name) = opt_curr_name {
+            if key_name == curr_name {
+                curr_types.insert(*key_rtype);
+            } else {
+                work.push((curr_name.clone(), curr_types));
+                opt_curr_name = Some(key_name);
+                curr_types = [*key_rtype].into();
+            }
+        } else {
+            opt_curr_name = Some(key_name);
+            curr_types.insert(*key_rtype);
+        }
+    }
+    if let Some(curr_name) = opt_curr_name {
+        work.push((curr_name.clone(), curr_types));
+    }
+    for (curr_name, mut curr_types) in work {
+        let mut curr_types = if curr_types.contains(&Rtype::NS) {
+            let has_ds = curr_types.contains(&Rtype::DS);
+            let mut curr_types: HashSet<Rtype> = [Rtype::NS].into();
+            if has_ds {
+                curr_types.insert(Rtype::DS);
+                curr_types.insert(Rtype::RRSIG);
+            }
+            curr_types
+        } else {
+            curr_types.insert(Rtype::RRSIG);
+            curr_types
+        };
+        let rtypebitmap = nsec3_rtypebitmap_from_iterator(curr_types.iter());
+
+        // Make sure NS doesn't get signed. And avoid signing RRSIGs.
+        curr_types.remove(&Rtype::NS);
+        curr_types.remove(&Rtype::RRSIG);
+        sign_rtype_set(&curr_name, &curr_types, iss)?;
+
+        let (nsec3_hash_octets, nsec3_name) = nsec3_hash_parts(&curr_name, iss);
+
+        nsec3_insert_full(&curr_name, nsec3_hash_octets, &nsec3_name, rtypebitmap, iss);
+    }
+    Ok(())
+}
+
+fn nsec3_insert_full(
+    name: &Name<Bytes>,
+    nsec3_hash: OwnerHash<Bytes>,
+    nsec3_name: &Name<Bytes>,
+    rtypebitmap: RtypeBitmap<Bytes>,
+    iss: &mut IncrementalSigningState,
+) {
+    nsec3_insert_one(nsec3_hash, nsec3_name, rtypebitmap, iss);
+
+    // Assume that we never insert the APEX. So the parent always exists.
+    let name = name.parent().expect("should exist");
+    nsec3_insert_ent(&name, iss);
+}
+
+fn nsec3_insert_ent(name: &Name<Bytes>, iss: &mut IncrementalSigningState) {
+    // Check if name has an NSEC3 record. If so, we are done. Otherwise,
+    // insert an ENT and continue with the parent.
+    let mut name = name.clone();
+    loop {
+        if !name.ends_with(&iss.origin) {
+            // This is weird, we should never be able to get beyond APEX.
+            // Just ignore this.
+            return;
+        }
+        if name == iss.origin {
+            // APEX exists by definition.
+            return;
+        }
+
+        let (nsec3_hash_octets, nsec3_name) = nsec3_hash_parts(&name, iss);
+
+        if iss.nsec3s.contains_key(&nsec3_name) {
+            // Found something. We are done.
+            return;
+        }
+
+        let rtypebitmap = RtypeBitmap::<Bytes>::builder();
+        let rtypebitmap = rtypebitmap.finalize();
+        nsec3_insert_one(nsec3_hash_octets, &nsec3_name, rtypebitmap, iss);
+
+        // Get the parent. We should be below APEX, so the parent has to exist.
+        name = name.parent().expect("parent should exist");
+    }
+}
+
+fn nsec3_insert_one(
+    nsec3_hash: OwnerHash<Bytes>,
+    nsec3_name: &Name<Bytes>,
+    rtypebitmap: RtypeBitmap<Bytes>,
+    iss: &mut IncrementalSigningState,
+) {
+    // Try to find the NSEC3 record that comes before the one we are trying
+    // to insert. It is possible that we try to insert before the first NSEC3
+    // record. In that case, logically try to insert after the last NSEC3
+    // record.
+    let mut range = iss.nsec3s.range::<Name<_>, _>(..nsec3_name);
+    let (previous_name, previous_record) = if let Some(kv) = range.next_back() {
+        kv
+    } else {
+        let mut range = iss.nsec3s.range::<Name<_>, _>(nsec3_name..);
+        range
+            .next_back()
+            .expect("at least one element should exist")
+    };
+    let previous_name = previous_name.clone();
+    let previous_record = previous_record.clone();
+    drop(range);
+    let ZoneRecordData::Nsec3(previous_nsec3) = previous_record.data() else {
+        panic!("NSEC3 record expected");
+    };
+    let next = previous_nsec3.next_owner();
+    let new_nsec3 = Nsec3::new(
+        iss.nsec3param.hash_algorithm(),
+        iss.nsec3param.flags(),
+        iss.nsec3param.iterations(),
+        iss.nsec3param.salt().clone(),
+        next.clone(),
+        rtypebitmap,
+    );
+    let new_record = Record::new(
+        nsec3_name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec3(new_nsec3),
+    );
+    iss.nsec3s.insert(nsec3_name.clone(), new_record);
+    iss.modified_nsecs.insert(nsec3_name.clone());
+    let previous_nsec3 = Nsec3::new(
+        iss.nsec3param.hash_algorithm(),
+        iss.nsec3param.flags(),
+        iss.nsec3param.iterations(),
+        iss.nsec3param.salt().clone(),
+        nsec3_hash,
+        previous_nsec3.types().clone(),
+    );
+    let previous_record = Record::new(
+        previous_name.clone(),
+        previous_record.class(),
+        previous_record.ttl(),
+        ZoneRecordData::Nsec3(previous_nsec3),
+    );
+    iss.nsec3s.insert(previous_name.clone(), previous_record);
+    iss.modified_nsecs.insert(previous_name.clone());
+}
+
+// Return the effective result HashSet even when the NSEC3 record gets deleted.
+fn nsec3_update_bitmap(
+    name: &Name<Bytes>,
+    nsec3_record: &Zrd,
+    nsec3: &Nsec3<Bytes>,
+    curr: &HashSet<Rtype>,
+    add: &HashSet<Rtype>,
+    delete: &HashSet<Rtype>,
+    iss: &mut IncrementalSigningState,
+) -> HashSet<Rtype> {
+    // Update curr.
+    let curr: HashSet<_> = curr.union(add).copied().collect();
+    let mut curr: HashSet<_> = curr.difference(delete).copied().collect();
+    let owner = nsec3_record.owner();
+
+    // Check if we need to add or remove RRSIG. Assume that apex has a SOA
+    // record.
+    if curr.contains(&Rtype::NS) && !curr.contains(&Rtype::SOA) {
+        // For an NS not at origin, there is an RRSIG if there is also a
+        // DS record.
+        if curr.contains(&Rtype::DS) {
+            // Yes, add RRSIG.
+            curr.insert(Rtype::RRSIG);
+        } else {
+            // No. Remove RRSIG.
+            curr.remove(&Rtype::RRSIG);
+        }
+    } else {
+        // Is there anything apart from RRSIG?
+        if curr.iter().any(|r| *r != Rtype::RRSIG) {
+            // Yes. Add RRSIG.
+            curr.insert(Rtype::RRSIG);
+        } else {
+            // No. Remove RRSIG.
+            curr.remove(&Rtype::RRSIG);
+        }
+    }
+
+    if curr.is_empty() {
+        // The NSEC3 bitmp will be empty, but this may now have become an
+        // empty non-terminal. Our only option is to update the NSEC3 record
+        // and then call nsec3_remove_et to see if it is empty can can be
+        // removed.
+        nsec3_update(owner, nsec3_record, nsec3, &curr, iss);
+        nsec3_remove_et(name, iss);
+        return curr;
+    }
+
+    if iss.nsec3param.opt_out_flag() && !curr.iter().any(|r| *r != Rtype::NS) {
+        // The new bitmap has nothing except for NS. We would like to delete
+        // the NSEC3. However there may still be descendents that need to be
+        // removed with nsec3_set_occluded. Update this NSEC3 to be empty and
+        // call nsec3_remove_et to remove it if there are no descendents.
+
+        let empty_curr = HashSet::new();
+        nsec3_update(owner, nsec3_record, nsec3, &empty_curr, iss);
+        nsec3_remove_et(name, iss);
+        return curr;
+    }
+
+    nsec3_update(owner, nsec3_record, nsec3, &curr, iss);
+    curr
+}
+
+fn nsec3_rtypebitmap_from_iterator<'a, I>(iter: I) -> RtypeBitmap<Bytes>
+where
+    I: Iterator<Item = &'a Rtype>,
+{
+    let mut rtypebitmap = RtypeBitmap::<Bytes>::builder();
+    for rtype in iter {
+        rtypebitmap.add(*rtype).expect("should not fail");
+    }
+    rtypebitmap.finalize()
+}
+
+fn nsec3_hash_parts(
+    name: &Name<Bytes>,
+    iss: &IncrementalSigningState,
+) -> (OwnerHash<Bytes>, Name<Bytes>) {
+    let nsec3_hash_octets = OwnerHash::<Bytes>::octets_from(
+        nsec3_hash::<_, _, BytesMut>(
+            name,
+            iss.nsec3param.hash_algorithm(),
+            iss.nsec3param.iterations(),
+            iss.nsec3param.salt(),
+        )
+        .expect("should not fail"),
+    );
+    let nsec3_hash_base32 = base32::encode_string_hex(&nsec3_hash_octets).to_ascii_lowercase();
+    let mut builder = NameBuilder::<BytesMut>::new();
+    builder
+        .append_label(nsec3_hash_base32.as_bytes())
+        .expect("should not fail");
+    let nsec3_name = builder.append_origin(&iss.origin).expect("should not fail");
+    (nsec3_hash_octets, nsec3_name)
+}
+
+fn is_occluded(name: &Name<Bytes>, iss: &IncrementalSigningState) -> bool {
+    // We need to check if the parent of name is a delegation. Stop
+    // when we reached origin.
+    let Some(mut curr) = name.parent() else {
+        // We asked for the parent of the root. That is weird. Just
+        // return not occluded.
+        return false;
+    };
+    loop {
+        if curr == iss.origin {
+            // We reached apex. The name was not occluded.
+            return false;
+        }
+        if !curr.ends_with(&iss.origin) {
+            // Something weird is going on. Return not occluded.
+            return false;
+        }
+        if iss.new_data.contains_key(&(curr.clone(), Rtype::NS)) {
+            // Name is occluded.
+            return true;
+        }
+        let Some(parent) = curr.parent() else {
+            // We asked for the parent of the root. That is weird. Just
+            // return not occluded.
+            return false;
+        };
+        curr = parent;
+    }
+}
+
+fn sign_rtype_set(
+    name: &Name<Bytes>,
+    set: &HashSet<Rtype>,
+    iss: &mut IncrementalSigningState,
+) -> Result<(), SignerError> {
+    let mut new_sigs = vec![];
+    for rtype in set {
+        let key = (name.clone(), *rtype);
+        let Some(records) = (if *name == iss.origin {
+            iss.new_apex.get(&key.1)
+        } else {
+            iss.new_data.get(&key)
+        }) else {
+            panic!("Expected something for {name}/{rtype}");
+        };
+        sign_records(
+            &iss.origin,
+            records,
+            &iss.keys,
+            iss.inception,
+            iss.expiration,
+            &mut new_sigs,
+        )?;
+    }
+    for (sig, rtype) in new_sigs {
+        let key = (sig[0].owner().clone(), rtype);
+        iss.rrsigs.insert(key, sig);
+    }
+    Ok(())
+}
